==> R/04_WDR/042_programs/02_data_cleaning_WDR.Rmd <==
---
title: "SPI Documentation for WDR Indicators"
author: "Brian Stacy"
date: "`r Sys.Date()`"
output: 
  html_document: 
    fig_height: 6
    fig_width: 8
    keep_md: yes
    toc: yes
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)

# Load Packages
library(tidyverse)
library(skimr)
library(kableExtra)
library(readxl)
library(Hmisc)
library(haven)
library(lubridate)
library(httr)
library(rsdmx)

# Directory for SPI csv files that were previously created
csv_dir <- "C:/Users/wb469649/OneDrive - WBG/Documents/Github/SPI_AKI/R/01_data/011_rawdata"

# Directory for SPI csv files that are saved
csv_output <- "C:/Users/wb469649/OneDrive - WBG/Documents/Github/SPI_AKI/R/04_WDR/043_outputs"

#get country metadata
country_list <- wbstats::wbcountries()


```

# Introduction

This file will clean the raw SPI data and form a set of indicators for use in the WDR.  

Hi Brian, 

Thanks so much for sending this my way – super super interesting materials. I am looking forward to seeing how the discussion with John Pullinger will evolve, and I’m sure I will be interested in citing some of the additional indicators. 
In the meantime, I have updated the excel document as discussed. It is quite a long list of requests, please don’t hesitate to let me know if this is too much of a burden on your time. I have put indicator breakdowns as income level (LIC, LMC, UMC, HIC) and WGB regions, though we probably won’t be citing regions very often so if it’s a lot of extra work, feel free to disregard that. On the issue of weighting, do you recommend population weighting or not? 

On the issue of child/infant mortality and maternal mortality, could you share with me the information as % of countries with vital reg source, survey source, no source as well as by each of the frequency intervals, broken down by income level? Again, please point out if too much work! 

On povcalnet, I am wondering what kind of information you have on comparability of the data over time? 

Again, thanks so much, Brian, for all your help with this, and please do let me know if this is asking too much!  
All the best,
Philip


## Table

| Section | Indicator | Source | SPI Inidicator | Proposed indicator |
|-|-|-|-|-|
| Coverage | coverage/availability score | ODIN | Could use SPI main score | Scores for dimensions 2 and 3 (separately), by income level and region |
| Coverage | Poverty since 2015 | Povcalnet | Also in SPI, but Povcalnet is   our source | See email |
| Coverage | Monthly or quarterly industrial production indices by income group | SCI | Quarterly GDP is included in SPI | % of countries with 3 or more values of quarterly GDP by incomelevel and   region; % of countries with no values in past 5 years by income level and   region |
| Coverage | Half of agricultural data needs met in SSA | GSARS | We have data on all agricultural   census or agriculture surveys in past 10 years | will drop this from text |
| Coverage | Gender statistics availability | UN Women | Gender indicator is Maternal   Mortality, but have been working to build database on whether breakdowns by   gender are available for several of first 10 SDGs | see email  |
| Coverage | PH Census in last 10 years | SCI | SPI as well, which we can   customize | % of countries with PH census in last ten years by income level and   region |
| Coverage | Ag Census in last 10 years | SCI | SPI as well, which we can   customize | same as PH census indicator |
| Coverage | Complete vital stats coverage | SCI  | SPI as well, which we can   customize | % of complete CRVS by income and region |
| Coverage | Number of people uncounted | UN Women | We do not have this | will use original source |
| Coverage | Geospatial: addresses | Global Open Data Index | We do not have this | will use original source |
| Coverage | Geospatial: maps | Global Open Data Index | We do not have this | will use original source |
| Coverage | Geospatial: national maps | Global Open Data Index | We do not have this | will use original source |
| Coverage | GDP per capita | SCI | We have actually dropped GDP per   capita, because we decided to include Quarterly GDP | use quarterly GDP, see above |
| Coverage | Child mortality | SCI | SPI has good child mortality   data | let's use, see email |
| Coverage | Immunization rates | SCI | Do not have | will drop this from text |
| Coverage | Disaggregation at admin1 | ODIN | Do not have | will use original source |
| Coverage | Disaggregation at admin2 | ODIN | Do not have | will use original source |
| Openness & Interoperability | Data Openness Score | ODIN | We have a subscore for data   dissemination practices and openness | Dimension 4 score by income level and region |
| Openness & Interoperability | Open terms of use agreement  | ODIN | Do not have | will be dropped or used from original source |
| Openness & Interoperability | Available in machine-readable format  | ODIN | Do not have | will be dropped or used from original source |
| Capacity | SCI score | SCI | We have this | (preliminary) SPI score by income level |
| Governance | NSO independence | Ibrahim Foundation | Do not have | will use original source |
| Additional   Indicators |  |  |  |  |
|  |  |  |  | 4.1, 4.2, 4.3, 4.5, 4.6, 4.7 by income level and region |
|  |  |  |  | Manufacturing, value added (% of GDP), % of countries without any in past   5 years by income and region |
|  |  |  |  | GNI current LCU, % of countries without any in past 5 years by income and   region |










# Indicator Cleaning

## SPI Dimension 1: Methodology, Standards and Classifications (MSC):


### 1.1 System of National Accounts in use

 The national accounts data are compiled using the concepts, definitions,
 framework, and methodology of the System of National Account 2008 (SNA2008)
 or European System of National and Regional Accounts (ESA 2010).  The manual
 has evolved to meet the changing economic structure, to follow systematic accounting
 and ensure international compatibility.

 Scoring: 1 point for using SNA2008 or ESA 2010, 0.5 points for using SNA 1993 or ESA 1995, 0 points otherwise

```{r snau}

#read in csv file.
D1.1.MSC.SNAU <- read_csv(file = paste(csv_dir, "D1.1.MSC.SNAU.csv", sep="/" )) %>%
  mutate(SPI.D1.1.SNAU=case_when(
    SNAU=="SNA 2008" ~ 1,
    SNAU=="SNA 1993" ~ 0.5,
    TRUE ~ 0
  )) %>%
  arrange(date, country) %>%
  select(iso3c, country, date,SNAU, SPI.D1.1.SNAU  )


D1.1.MSC.SNAU_wide <- D1.1.MSC.SNAU %>%
  pivot_wider(names_from=date,
              names_sep = "_",
              values_from=c('SNAU', 'SPI.D1.1.SNAU')
              ) %>%
  left_join(country_list)

write_excel_csv(D1.1.MSC.SNAU_wide,
                path = paste(csv_output, "D1.1.MSC.SNAU_indicator.csv", sep="/" ))
```


### 1.2 National Accounts base year

National accounts base year is the year used as the base period for constant price
calculations in the country's national accounts.  It is recommended that the base year of
constant price estimates be changed periodically to reflect changes in economic structure and relative prices.

1 point for chained price, 0.5 for reference period within past 10 years, 0 points otherwise.

```{r naby}

#read in csv file.
D1.2.MSC.NABY <- read_csv(file = paste(csv_dir, "D1.2.MSC.NABY.csv", sep="/" )) %>%
  mutate(SPI.D1.2.NABY=case_when(
    NABY=="Original chained constant price data are rescaled." ~ 1,
    (date-as.numeric(NABY))<=10 ~ 0.5, #within 10 years of reference period
    TRUE ~ 0
  ))  %>%
  select(iso3c, country, date, NABY, SPI.D1.2.NABY  ) %>%
  arrange(date, country)

D1.2.MSC.NABY_wide <- D1.2.MSC.NABY %>%
  pivot_wider(names_from=date,
              names_sep = "_",
              values_from=c('NABY', 'SPI.D1.2.NABY')
              ) %>%
  left_join(country_list)

write_excel_csv(D1.2.MSC.NABY_wide,
                path = paste(csv_output, "D1.2.MSC.NABY_indicator.csv", sep="/" ))

```

### 1.3 Classification of national industry

 The industrial production data are compiled using the International Standard Industrial
 Classification of All Economic Activities (ISIC) Rev.4 and Statistical Classification of
 Economic Activities in the European Community (NACE) Rev.2.  ISIC Rev.4 is a standard classification
 of economic activities arranged so that entities can be classified per the activity they carry out
 using criteria such as input, output and use of the products produced, more emphasis has been given
 to the character of the production process in defining and delineating ISIC classes for international
 comparability.  The manual and classification have changed to cover the complete scope of industrial
 production, employment, and GDP and other statistical areas.

1 Point. Latest version is adopted (ISIC Rev 4, NACE Rev 2 or a compatible classification)

0.5 Points.  Previous version is used (ISIC Rev 3, NACE Rev 1 or a compatible classification)

0 Points. Otherwise

```{r cnin}

#read in csv file.
D1.3.MSC.CNIN <- read_csv(file = paste(csv_dir, "D1.3.MSC.CNIN.csv", sep="/" )) %>%
  mutate(SPI.D1.3.CNIN=case_when(
    str_to_lower(CNIN)=="nace rev2" | str_to_lower(CNIN)=="rev4" ~ 1,
    str_to_lower(CNIN)=="nace rev1" | str_to_lower(CNIN)=="rev3" ~ 0.5,
    TRUE ~ 0
  ))  %>%
  select(iso3c, country, date, CNIN, SPI.D1.3.CNIN  ) %>%
  arrange(date, country)

D1.3.MSC.CNIN_wide <- D1.3.MSC.CNIN %>%
  pivot_wider(names_from=date,
              names_sep = "_",
              values_from=c('CNIN', 'SPI.D1.3.CNIN')
              ) %>%
  left_join(country_list)

write_excel_csv(D1.3.MSC.CNIN_wide,
                path = paste(csv_output, "D1.3.MSC.CNIN_indicator.csv", sep="/" ))
```

### 1.4 CPI base year

Consumer Price Index serves as indicators of inflation and reflects changes in the
 cost of acquiring a fixed basket of goods and services by the average consumer.
 Weights are usually derived from consumer expenditure surveys and the CPI base year
 refers to the year the weights were derived.  It is recommended that the base year be
 changed periodically to reflect changes in expenditure structure.

 1 Point. Annual chain linking. 0.5 Points. Base year in last 10 years. 0 points. Otherwise

```{r cpiby}

#read in csv file.
D1.4.MSC.CPIBY <- read_csv(file = paste(csv_dir, "D1.4.MSC.CPIBY.csv", sep="/" )) %>%
  mutate(SPI.D1.4.CPIBY=case_when(
    CPIBY=="annual chained" ~ 1,
    (date-as.numeric(CPIBY))<=10 ~ 0.5, #within 10 years of reference period
    TRUE ~ 0
  ))  %>%
  select(iso3c, country, date, CPIBY, SPI.D1.4.CPIBY  ) %>%
  arrange(date, country)

D1.4.MSC.CPIBY_wide <- D1.4.MSC.CPIBY %>%
  pivot_wider(names_from=date,
              names_sep = "_",
              values_from=c('CPIBY', 'SPI.D1.4.CPIBY')
              ) %>%
  left_join(country_list)

write_excel_csv(D1.4.MSC.CPIBY_wide,
                path = paste(csv_output, "D1.4.MSC.CPIBY_indicator.csv", sep="/" ))
```




### 1.5 Classification of household consumption

Classification of Individual Consumption According to Purpose (COICOP)
is used in household budget surveys, consumer price indices and international
comparisons of gross domestic product (GDP) and its component expenditures.
Although COICOP is not strictly linked to any particular model of consumer
behavior, the classification is designed to broadly reflect differences in
income elasticities.  It is an integral part of the SNA1993 and more detailed
subdivision of the classes provide comparability between countries and between
statistics in these different areas.

1 Point. Follow Classification of Individual Consumption by Purpose (COICOP)	0 Points.	Otherwise

```{r hous}

#read in csv file.
D1.5.MSC.HOUS <- read_csv(file = paste(csv_dir, "D1.5.MSC.HOUS.csv", sep="/" )) %>%
  mutate(SPI.D1.5.HOUS=case_when(
    HOUS=="COICOP" ~ 1,
    TRUE ~ 0
  ))  %>%
  select(iso3c, country, date, HOUS, SPI.D1.5.HOUS  ) %>%
  arrange(date, country)

D1.5.MSC.HOUS_wide <- D1.5.MSC.HOUS %>%
  pivot_wider(names_from=date,
              names_sep = "_",
              values_from=c('HOUS', 'SPI.D1.5.HOUS')
              ) %>%
  left_join(country_list)

write_excel_csv(D1.5.MSC.HOUS_wide,
                path = paste(csv_output, "D1.5.MSC.HOUS_indicator.csv", sep="/" ))
```




### 1.6 Classification of status of employment

Classification of status of employment refers to employment data that are
 compiled using the current international standard International Classification
 of Status in Employment (ISCE-93).  It classifies jobs with respect to the type
 of explicit or implicit contract of employment between the job holder and the
 economic unit in which he or she is employed.  Therefore, it aims to provide the
 basis for production of internationally comparable statistics on the employment
 relationship, including the distinction between salaried employment and self-employment.

1 Point. Follow International Labour Organization, International Classification of Status in Employment (ICSE-93) or 2012 North American Industry Classification System (NAICS). 0 Points Otherwise.

```{r empl}

#read in csv file.
D1.6.MSC.EMPL <- read_csv(file = paste(csv_dir, "D1.6.MSC.EMPL.csv", sep="/" )) %>%
  mutate(SPI.D1.6.EMPL=case_when(
    EMPL=="ICSE-93" | EMPL=="NAICS" ~ 1,
    TRUE ~ 0
  ))  %>%
  select(iso3c, country, date, EMPL, SPI.D1.6.EMPL  ) %>%
  arrange(date, country)

D1.6.MSC.EMPL_wide <- D1.6.MSC.EMPL %>%
  pivot_wider(names_from=date,
              names_sep = "_",
              values_from=c('EMPL', 'SPI.D1.6.EMPL')
              ) %>%
  left_join(country_list)

write_excel_csv(D1.6.MSC.EMPL_wide,
                path = paste(csv_output, "D1.6.MSC.EMPL_indicator.csv", sep="/" ))
```



### 1.7 Central government accounting status

Government finance accounting status refers to the accounting basis for
reporting central government financial data.  For many countries' government
finance data, have been consolidated into one set of accounts capturing all
the central government's fiscal activities and following noncash recording basis.
Budgetary central government accounts do not necessarily include all central government
units, the picture they provide of central government activities is usually incomplete.

1 Point. Consolidated central government accounting follows noncash recording basis
0.5 Points. Consolidated central government accounting follows cash recording basis
0 Points. Otherwise

```{r cgov}

#read in csv file.
D1.7.MSC.CGOV <- read_csv(file = paste(csv_dir, "D1.7.MSC.CGOV.csv", sep="/" )) %>%
  mutate(SPI.D1.7.CGOV=case_when(
    CGOV=="AC" ~ 1,
    CGOV=="CA" ~ 0.5,
    TRUE ~ 0
  ))  %>%
  select(iso3c, country, date, CGOV, SPI.D1.7.CGOV  ) %>%
  arrange(date, country)

D1.7.MSC.CGOV_wide <- D1.7.MSC.CGOV %>%
  pivot_wider(names_from=date,
              names_sep = "_",
              values_from=c('CGOV', 'SPI.D1.7.CGOV')
              ) %>%
  left_join(country_list)

write_excel_csv(D1.7.MSC.CGOV_wide,
                path = paste(csv_output, "D1.7.MSC.CGOV_indicator.csv", sep="/" ))
```

### 1.8 Compilation of government finance statistics

(GFSM) in use for compiling the data.  It provides guidelines on the institutional
structure of governments and the presentation of fiscal data in a format similar to
business accounting with a balance sheet and income statement plus guidelines on the
treatment of exchange rate and other valuation adjustments.  The latest manual GFSM2014
is harmonized with the SNA2008.

1 Point. Follow the latest Government Finance Statistical Manual (2014)/ ESA2010
0.5 Points. Previous version is used (GFSM 2001)
0 Points. Otherwise


```{r fina}

#read in csv file.
D1.8.MSC.FINA <- read_csv(file = paste(csv_dir, "D1.8.MSC.FINA.csv", sep="/" )) %>%
  mutate(SPI.D1.8.FINA=case_when(
    FINA=="2014" | FINA=="ESA 2010"~ 1,
    FINA=="2001" ~ 0.5,
    TRUE ~ 0
  ))  %>%
  select(iso3c, country, date, FINA, SPI.D1.8.FINA  ) %>%
  arrange(date, country)

D1.8.MSC.FINA_wide <- D1.8.MSC.FINA %>%
  pivot_wider(names_from=date,
              names_sep = "_",
              values_from=c('FINA', 'SPI.D1.8.FINA')
              ) %>%
  left_join(country_list)

write_excel_csv(D1.8.MSC.FINA_wide,
                path = paste(csv_output, "D1.8.MSC.FINA_indicator.csv", sep="/" ))
```



### 1.9 Compilation of monetary and financial statistics

Compilation of monetary and financial statistics refers to the Monetary and Financial Statistics Manual
(MFSM) in use.  It covers concepts, definitions, classifications of financial instruments and sectors, and
accounting rules, and provides a comprehensive analytic framework for monetary and financial planning and policy
determination.  The Monetary and Finance Statistics: Compilation Guide (2008) provides detailed guidelines for
the compilation of monetary and financial statistics in addition to MFSM.

1 Point. Follow the latest Monetary and Finance Statistics Manual (2000) or Monetary and Finance Statistics: Compilation Guide (2008/2016)
0 Points. Otherwise

```{r mony}

#read in csv file.
D1.9.MSC.MONY <- read_csv(file = paste(csv_dir, "D1.9.MSC.MONY.csv", sep="/" )) %>%
  mutate(SPI.D1.9.MONY=case_when(
    MONY=="MFSM 2000" | MONY=="MFSMCG 2016"~ 1,
    TRUE ~ 0
  ))  %>%
  select(iso3c, country, date, MONY, SPI.D1.9.MONY  ) %>%
  arrange(date, country)

D1.9.MSC.MONY_wide <- D1.9.MSC.MONY %>%
  pivot_wider(names_from=date,
              names_sep = "_",
              values_from=c('MONY', 'SPI.D1.9.MONY')
              ) %>%
  left_join(country_list)

write_excel_csv(D1.9.MSC.MONY_wide,
                path = paste(csv_output, "D1.9.MSC.MONY_indicator.csv", sep="/" ))
```



### 1.10 SDDS/e-GDDS subscription

Data Dissemination Standard (SDDS) and electronic General Data Dissemination Standard (e-GDDS) were
established by the International Monetary Fund (IMF) for member countries that have or that might seek
access to international capital markets, to guide them in providing their economic and financial data to the public.
Although subscription is voluntary, the subscribing member needs to be committed to observing the standard and provide
information about its data and data dissemination practices (metadata).  The metadata are posted on the IMF's SDDS and
e-GDDS websites.

1 Point. Subscribing to IMF SDDS+ or SDDS standards
0.5 Points. Subscribing to IMF e-GDDS standards
0 Points. Otherwise

```{r idds}

#read in csv file.
D1.10.MSC.IDDS <- read_csv(file = paste(csv_dir, "D1.10.MSC.IDDS.csv", sep="/" )) %>%
  mutate(SPI.D1.10.IDDS=case_when(
    IDDS=="SDDS Plus" | IDDS=="SDDS"~ 1,
    IDDS=="e-GDDS"~ 0.5,
    TRUE ~ 0
  ))  %>%
  select(iso3c, country, date, IDDS, SPI.D1.10.IDDS  ) %>%
  arrange(date, country)

D1.10.MSC.IDDS_wide <- D1.10.MSC.IDDS %>%
  pivot_wider(names_from=date,
              names_sep = "_",
              values_from=c('IDDS', 'SPI.D1.10.IDDS')
              ) %>%
  left_join(country_list)

write_excel_csv(D1.10.MSC.IDDS_wide,
                path = paste(csv_output, "D1.10.MSC.IDDS_indicator.csv", sep="/" ))
```



### 1.11 CRVS

Civil registration and vital statistics record the occurrence and characteristics of
vital events (births, deaths, marriage and divorce etc.) pertaining to the population and
serve as a main source of vital statistics. This identifies countries that report at least
90 percent complete registries of vital (birth and death) statistics to the United Nations
Statistics Division and are reported in its Population and Vital Statistics Reports.
Countries with complete vital statistics registries may have more accurate and timely demographic indicators.

1 Point. Vital registration complete
0 Points. Otherwise
```{r crvs}

#read in csv file.
D1.11.MSC.CRVS <- read_csv(file = paste(csv_dir, "D1.11.MSC.CRVS.csv", sep="/" )) %>%
  mutate(SPI.D1.11.CRVS=case_when(
    str_to_lower(CRVS)=="yes" ~ 1,
    TRUE ~ 0
  ))  %>%
  select(iso3c, country, date, CRVS, SPI.D1.11.CRVS  ) %>%
  arrange(date, country)

D1.11.MSC.CRVS_wide <- D1.11.MSC.CRVS %>%
  pivot_wider(names_from=date,
              names_sep = "_",
              values_from=c('CRVS', 'SPI.D1.11.CRVS')
              ) %>%
  left_join(country_list)

write_excel_csv(D1.11.MSC.CRVS_wide,
                path = paste(csv_output, "D1.11.MSC.CRVS_indicator.csv", sep="/" ))
```



### 1.12 Business process

The Generic Statistical Business Process Model (GSBPM) aims to describe
statistics production in a general and process-oriented way.  It is used
both within and between statistical offices as a common basis for work with
statistics production in different ways, such as quality, efficiency, standardization,
and process-orientation.  It is used for all types of surveys, and "business" is not
related to "business statistics" but refers to the statistical office, simply expressed.

1 Point. GSBPM is in use
0 Points. Otherwise

```{r gsbp}

#read in csv file.
D1.12.MSC.GSBP <- read_csv(file = paste(csv_dir, "D1.12.MSC.GSBP.csv", sep="/" )) %>%
  mutate(SPI.D1.12.GSBP=case_when(
    str_to_lower(GSBP)=="yes" ~ 1,
    TRUE ~ 0
  ))  %>%
  select(iso3c, country, date, GSBP, SPI.D1.12.GSBP  ) %>%
  arrange(date, country)

D1.12.MSC.GSBP_wide <- D1.12.MSC.GSBP %>%
  pivot_wider(names_from=date,
              names_sep = "_",
              values_from=c('GSBP', 'SPI.D1.12.GSBP')
              ) %>%
  left_join(country_list)

write_excel_csv(D1.12.MSC.GSBP_wide,
                path = paste(csv_output, "D1.12.MSC.GSBP_indicator.csv", sep="/" ))
```

### MSC Score

MSC Country  Score=Weighted Score÷Maximum Category Score×100

```{r msc}
# join datasets together based on country and date
D1.MSC <- D1.1.MSC.SNAU %>%
  left_join(D1.2.MSC.NABY) %>%
  left_join(D1.3.MSC.CNIN) %>%
  left_join(D1.4.MSC.CPIBY) %>%
  left_join(D1.5.MSC.HOUS) %>%
  left_join(D1.6.MSC.EMPL) %>%
  left_join(D1.7.MSC.CGOV) %>%
  left_join(D1.8.MSC.FINA) %>%
  left_join(D1.9.MSC.MONY) %>%
  left_join(D1.10.MSC.IDDS) %>%
  left_join(D1.11.MSC.CRVS) %>%
  left_join(D1.12.MSC.GSBP) %>%
  select(iso3c, country, date, starts_with("SPI.D1"))
#Now calculate MSC score which is the average across the 12 indicators
D1.MSC <- D1.MSC %>%
  mutate(SPI.D1.MSC=100*rowMeans(.[grep(x=colnames(D1.MSC),
                                              pattern="SPI.D1")], na.rm=TRUE)) %>%
  arrange(date, iso3c) %>%
  select(iso3c, country, date,SPI.D1.MSC, starts_with("SPI.D1"))


D1.MSC_wide <- D1.MSC %>%
  select(iso3c, country, date,SPI.D1.MSC) %>%
  pivot_wider(names_from=date,
              names_prefix = "SPI.D1.MSC_",
              values_from=c('SPI.D1.MSC')
              ) %>%
  left_join(country_list)

write_excel_csv(D1.MSC_wide,
                path = paste(csv_output, "D1.MSC_indicator.csv", sep="/" ))

```

Produce summary statistics of our indicators.

```{r}

msc_sumstats <- D1.MSC %>%
  group_by(date) %>%
  select(SPI.D1.MSC, starts_with("SPI.D1")) %>%
  skim() %>%
  select(c("skim_variable",
           "date", "complete_rate", "numeric.mean",  "numeric.sd",    "numeric.p0",    "numeric.p25",   "numeric.p50",
           "numeric.p75",   "numeric.p100",  "numeric.hist"))

kable(msc_sumstats, caption="Summary Statistics of Methodology, Standards and Classifications Indicator Variables" , col.names = c("Indicator", "Date", "Completion Rate", "Mean",  "Std Dev","Min", "25th Percentile", "Median", "75th Percentile", "Max",  "Histogram"), digits = 2 ) %>%
  kable_styling(bootstrap_options = c("striped", "hover"))

```


## SPI Dimension 2: Censuses and Surveys (CS): 

### 2.1 Population & Housing census

Population censuses collect data on the size, distribution and composition of population and 
information on a broad range of social and economic characteristics of the population.  It also 
provides sampling frames for household and other surveys.  Housing censuses provide information 
on the supply of housing units, the structural characteristics and facilities, and health and the 
development of normal family living conditions.  Data obtained as part of the population census, 
including data on homeless persons, are often used in the presentation and analysis of the results 
of the housing census.  It is recommended that population and housing censuses be conducted at least every 10 years. 

1 Point. Population census done within last 10 years	
0 Points. Otherwise
```{r popu}

#read in csv file.
D2.1.CEN.POPU_df <- read_csv(file = paste(csv_dir, "D2.1.CEN.POPU.csv", sep="/" )) %>%
  group_by(iso3c, country) %>% 
  nest() %>% # The next chunk of code will split our string with the years of the census (i.e. "2000, 2010") in to separate rows.  We will then aggregate up.
  mutate(
    temp_col = map(
      data, 
      ~ str_extract_all(.x$POPU.CENSUS, "\\d{4}") %>% 
        flatten() %>% 
        map_chr(~return(.x)) %>% 
        as_tibble()
    )
  ) %>% 
  unnest(keep_empty = TRUE) %>% # Now we have a database with the observations equal to Country*Census observations.  From here we can calculate latest census, etc.
  mutate(census_date=as.numeric(value)) 



# Now calculate our SPI score for this indicator
for (i in 2016:2019) {
 temp <- D2.1.CEN.POPU_df %>%
  mutate(date=i) %>%
  mutate(recency_indicator=((date>census_date)) ) %>% #restrict to censuses that do not occur after reference date
  mutate(census_date=if_else(recency_indicator==TRUE, census_date, as.numeric(NA))) %>%
  group_by(iso3c, country, date, database_last_updated, POPU.CENSUS ) %>%   
  summarise(last_census=max(census_date, na.rm=T)) %>%
  mutate(SPI.D2.1.POPU=case_when(
    (date-last_census<=10) & (date-last_census>0) ~ 1, 
    #(date-last_census<=20) & (date-last_census>0) ~ 0.5, 
    TRUE ~ 0 )
  )  %>%
   ungroup() %>%
  select(iso3c, country, date, POPU.CENSUS, SPI.D2.1.POPU  ) %>%
  arrange(date, country)

  
  assign(paste("D2.1.CEN.POPU",i,sep="_"), temp)
}


D2.1.CEN.POPU<- bind_rows(D2.1.CEN.POPU_2016, D2.1.CEN.POPU_2017, D2.1.CEN.POPU_2018, D2.1.CEN.POPU_2019)

D2.1.CEN.POPU_wide <- D2.1.CEN.POPU %>%
  pivot_wider(names_from=date,
              names_sep = "_",
              values_from=c('POPU.CENSUS', 'SPI.D2.1.POPU')
              ) %>%
  left_join(country_list)

write_excel_csv(D2.1.CEN.POPU_wide,
                path = paste(csv_output, "D2.1.CEN.POPU_indicator.csv", sep="/" ))
```


### 2.2 Agriculture census

Agriculture censuses collect information on agricultural activities, such as
size of holding, land tenure, land use, employment and production, and provide
basic structural data and sampling frames for agricultural surveys.  Censuses
of agriculture normally involves collecting key structural data by complete
enumeration of all agricultural holdings, in combination with more detailed
structural data using sampling methods.  It is recommended that agricultural
censuses be conducted at least every 10 years.

1 Point.  census done within last 10 years	
0 Points. Otherwise

```{r agri}

#read in csv file.
D2.2.CEN.AGRI_df <- read_csv(file = paste(csv_dir, "D2.2.CEN.AGRI.csv", sep="/" )) %>%
  group_by(iso3c, country) %>% 
  nest() %>% # The next chunk of code will split our string with the years of the census (i.e. "2000, 2010") in to separate rows.  We will then aggregate up.
  mutate(
    temp_col = map(
      data, 
      ~ str_extract_all(.x$AGRI.CENSUS, "\\d{4}") %>% 
        flatten() %>% 
        map_chr(~return(.x)) %>% 
        as_tibble()
    )
  ) %>% 
  unnest(keep_empty = TRUE) %>% # Now we have a database with the observations equal to Country*Census observations.  From here we can calculate latest census, etc.
  mutate(census_date=as.numeric(value)) 


# Now calculate our SPI score for this indicator
for (i in 2016:2019) {
 temp <- D2.2.CEN.AGRI_df %>%
  mutate(date=i) %>%
  mutate(recency_indicator=((date>census_date)) ) %>% #restrict to censuses that do not occur after reference date
  mutate(census_date=if_else(recency_indicator==TRUE, census_date, as.numeric(NA))) %>%
  group_by(iso3c, country, date, database_last_updated, AGRI.CENSUS ) %>%   
  summarise(last_census=max(census_date, na.rm=T)) %>%
  mutate(SPI.D2.2.AGRI=case_when(
    (date-last_census<=10) & (date-last_census>0) ~ 1, 
    #(date-last_census<=20) & (date-last_census>0) ~ 0.5, 
    TRUE ~ 0 )
  )  %>%
   ungroup() %>%
  select(iso3c, country, date, AGRI.CENSUS, SPI.D2.2.AGRI  ) %>%
  arrange(date, country)

  
  assign(paste("D2.2.CEN.AGRI",i,sep="_"), temp)
}


D2.2.CEN.AGRI<- bind_rows(D2.2.CEN.AGRI_2016, D2.2.CEN.AGRI_2017, D2.2.CEN.AGRI_2018, D2.2.CEN.AGRI_2019)

D2.2.CEN.AGRI_wide <- D2.2.CEN.AGRI %>%
  pivot_wider(names_from=date,
              names_sep = "_",
              values_from=c('AGRI.CENSUS', 'SPI.D2.2.AGRI')
              ) %>%
  left_join(country_list)

write_excel_csv(D2.2.CEN.AGRI_wide,
                path = paste(csv_output, "D2.2.CEN.AGRI_indicator.csv", sep="/" ))
```

### 2.3 Business/establishment census

Business/establishment censuses provide valuable information on all economic
activities, number of employed and size of establishments in the economy.
Business Register information is establishment-based and includes business
location, organization type (e.g. subsidiary or parent), industry
classification, and operating data (e.g., receipts and employment).

1 Point.  census done within last 10 years
0 Points. Otherwise


```{r cen.bizz}

#read in csv file.
D2.3.CEN.BIZZ_df <- read_csv(file = paste(csv_dir, "D2.3.CEN.BIZZ.csv", sep="/" )) %>%
  group_by(iso3c, country) %>%
  nest() %>% # The next chunk of code will split our string with the years of the census (i.e. "2000, 2010") in to separate rows.  We will then aggregate up.
  mutate(
    temp_col = map(
      data,
      ~ str_extract_all(.x$BIZZ.CENSUS, "\\d{4}") %>%
        flatten() %>%
        map_chr(~return(.x)) %>%
        as_tibble()
    )
  ) %>%
  unnest(keep_empty = TRUE) %>% # Now we have a database with the observations equal to Country*Census observations.  From here we can calculate latest census, etc.
  mutate(census_date=as.numeric(value))



# Now calculate our SPI score for this indicator
for (i in 2016:2019) {
 temp <- D2.3.CEN.BIZZ_df %>%
  mutate(date=i) %>%
  mutate(recency_indicator=((date>census_date)) ) %>% #restrict to censuses that do not occur after reference date
  mutate(census_date=if_else(recency_indicator==TRUE, census_date, as.numeric(NA))) %>%
  group_by(iso3c, country, date, database_last_updated, BIZZ.CENSUS ) %>%
  summarise(last_census=max(census_date, na.rm=T)) %>%
  mutate(SPI.D2.3.BIZZ=case_when(
    (date-last_census<=10) & (date-last_census>0) ~ 1,
    #(date-last_census<=20) & (date-last_census>0) ~ 0.5,
    TRUE ~ 0 )
  )  %>%
   ungroup() %>%
  select(iso3c, country, date, BIZZ.CENSUS, SPI.D2.3.BIZZ  ) %>%
  arrange(date, country)


  assign(paste("D2.3.CEN.BIZZ",i,sep="_"), temp)
}


D2.3.CEN.BIZZ<- bind_rows(D2.3.CEN.BIZZ_2016, D2.3.CEN.BIZZ_2017, D2.3.CEN.BIZZ_2018, D2.3.CEN.BIZZ_2019)

D2.3.CEN.BIZZ_wide <- D2.3.CEN.BIZZ %>%
  pivot_wider(names_from=date,
              names_sep = "_",
              values_from=c('BIZZ.CENSUS', 'SPI.D2.3.BIZZ')
              ) %>%
  left_join(country_list)

write_excel_csv(D2.3.CEN.BIZZ_wide,
                path = paste(csv_output, "D2.3.CEN.BIZZ_indicator.csv", sep="/" ))
```


### 2.4 Household Survey on income/consumption/expenditure/budget/Integrated Survey

These surveys collect data on household income (including income in kind),
consumption and expenditure.  They typically include income, expenditure, and
consumption surveys, household budget surveys, integrated surveys.  It is
recommended that surveys on income and expenditure be conducted at least every
3 to 5 years.

1 Point. 3 or more surveys done within past 10 years

0 Points. None within past 10 years

```{r svy.hous}

#read in csv file.
D2.4.SVY.HOUS_df <- read_csv(file = paste(csv_dir, "D2.4.SVY.HOUS.csv", sep="/" )) %>%
  group_by(iso3c, country) %>%  # The next chunk of code will split our string with the years of the survey (i.e. "2000, 2010") in to separate rows.  We will then aggregate up.
  nest() %>%
  mutate(
    temp_col = map(
      data,
      ~ str_extract_all(.x$HOUS.SURVEYS, "\\d{4}") %>%
        flatten() %>%
        map_chr(~return(.x)) %>%
        as_tibble()
    )
  ) %>%
  unnest(keep_empty = TRUE) %>% # Now we have a database with the observations equal to Country*Census observations.  From here we can calculate latest census, etc.
  mutate(survey_date=as.numeric(value))


# Now calculate our SPI score for this indicator
for (i in 2016:2019) {
 temp <- D2.4.SVY.HOUS_df %>%
  mutate(date=i) %>%
  mutate(recency_indicator=((date-survey_date<=10) & (date-survey_date>0)) ) %>% #restrict to surveys inside 10 year period
  group_by(iso3c, country, date, HOUS.SURVEYS) %>% #group by country and create indicator for how many surveys over 10 year period
  summarise(SPI.D2.4.HOUS=case_when(
    sum(recency_indicator)>=3 ~ 1,
    # sum(recency_indicator)==2 ~ 0.6,
    # sum(recency_indicator)==1 ~ 0.3,
    TRUE ~ 0 )
  )  %>%
  select(iso3c, country, date, HOUS.SURVEYS, SPI.D2.4.HOUS  ) %>%
  arrange(date, country)


  assign(paste("D2.4.SVY.HOUS",i,sep="_"), temp)
}

D2.4.SVY.HOUS<- bind_rows(D2.4.SVY.HOUS_2016, D2.4.SVY.HOUS_2017, D2.4.SVY.HOUS_2018, D2.4.SVY.HOUS_2019)


D2.4.SVY.HOUS_wide <- D2.4.SVY.HOUS %>%
  pivot_wider(names_from=date,
              names_sep = "_",
              values_from=c('HOUS.SURVEYS', 'SPI.D2.4.HOUS')
              ) %>%
  left_join(country_list)

write_excel_csv(D2.4.SVY.HOUS_wide,
                path = paste(csv_output, "D2.4.SVY.HOUS_indicator.csv", sep="/" ))
```




### 2.5 Agriculture survey

Agricultural surveys refer to surveys of agricultural holdings based on the
sampling frames established by the agricultural census.  These are surveys on
agricultural land, production, crops and livestock, aquaculture, labor and
cost, and time use.  Some issues, such as gender and food security, are of
interest to most agriculture surveys.

1 Point. 3 or more surveys done within past 10 years

0 Points. None within past 10 years


```{r svy.agri}

#read in csv file.
D2.5.SVY.AGRI_df <- read_csv(file = paste(csv_dir, "D2.5.SVY.AGRI.csv", sep="/" )) %>%
  group_by(iso3c, country) %>%  # The next chunk of code will split our string with the years of the survey (i.e. "2000, 2010") in to separate rows.  We will then aggregate up.
  nest() %>%
  mutate(
    temp_col = map(
      data,
      ~ str_extract_all(.x$AGRI.SURVEYS, "\\d{4}") %>%
        flatten() %>%
        map_chr(~return(.x)) %>%
        as_tibble()
    )
  ) %>%
  unnest(keep_empty = TRUE) %>% # Now we have a database with the observations equal to Country*Census observations.  From here we can calculate latest census, etc.
  mutate(survey_date=as.numeric(value))


# Now calculate our SPI score for this indicator
for (i in 2016:2019) {
 temp <- D2.5.SVY.AGRI_df %>%
  mutate(date=i) %>%
  mutate(recency_indicator=((date-survey_date<=10) & (date-survey_date>0)) ) %>% #restrict to surveys inside 10 year period
  group_by(iso3c, country, date, AGRI.SURVEYS) %>% #group by country and create indicator for how many surveys over 10 year period
  summarise(SPI.D2.5.AGRI=case_when(
    sum(recency_indicator)>=3 ~ 1,
   # sum(recency_indicator)==2 ~ 0.6,
   # sum(recency_indicator)==1 ~ 0.3,
    TRUE ~ 0 )
  )  %>%
  select(iso3c, country, date, AGRI.SURVEYS, SPI.D2.5.AGRI  ) %>%
  arrange(date, country)


  assign(paste("D2.5.SVY.AGRI",i,sep="_"), temp)
}

D2.5.SVY.AGRI<- bind_rows(D2.5.SVY.AGRI_2016, D2.5.SVY.AGRI_2017, D2.5.SVY.AGRI_2018,D2.5.SVY.AGRI_2019)

D2.5.SVY.AGRI_wide <- D2.5.SVY.AGRI %>%
  pivot_wider(names_from=date,
              names_sep = "_",
              values_from=c('AGRI.SURVEYS', 'SPI.D2.5.AGRI')
              ) %>%
  left_join(country_list) 

write_excel_csv(D2.5.SVY.AGRI_wide,
                path = paste(csv_output, "D2.5.SVY.AGRI_indicator.csv", sep="/" ))
```



### 2.6 Labor Force Survey

Labor force survey is a standard household-based survey of work-related
statistics at the national and sub-national employment or unemployment levels,
rates or trends.  The surveys also provide the characteristics of the employed
or unemployed, including labor force status by age or gender, breakdowns
between employees and the self-employed, public versus private sector
employment, multiple job-holding, hiring, job creation, and duration of
unemployment.

1 Point. 3 or more surveys done within past 10 years

0 Points. None within past 10 years



```{r svy.labr}

#read in csv file.
D2.6.SVY.LABR_df <- read_csv(file = paste(csv_dir, "D2.6.SVY.LABR.csv", sep="/" )) %>%
  group_by(iso3c, country) %>%  # The next chunk of code will split our string with the years of the survey (i.e. "2000, 2010") in to separate rows.  We will then aggregate up.
  nest() %>%
  mutate(
    temp_col = map(
      data,
      ~ str_extract_all(.x$LABR.SURVEYS, "\\d{4}") %>%
        flatten() %>%
        map_chr(~return(.x)) %>%
        as_tibble()
    )
  ) %>%
  unnest(keep_empty = TRUE) %>% # Now we have a database with the observations equal to Country*Census observations.  From here we can calculate latest census, etc.
  mutate(survey_date=as.numeric(value))

# Now calculate our SPI score for this indicator
for (i in 2016:2019) {
 temp <- D2.6.SVY.LABR_df %>%
  mutate(date=i) %>%
  mutate(recency_indicator=((date-survey_date<=10) & (date-survey_date>0)) ) %>% #restrict to surveys inside 10 year period
  group_by(iso3c, country, date, LABR.SURVEYS) %>% #group by country and create indicator for how many surveys over 10 year period
  summarise(SPI.D2.6.LABR=case_when(
    sum(recency_indicator)>=3 ~ 1,
   # sum(recency_indicator)==2 ~ 0.6,
   # sum(recency_indicator)==1 ~ 0.3,
    TRUE ~ 0 )
  )  %>%
  select(iso3c, country, date, LABR.SURVEYS, SPI.D2.6.LABR  ) %>%
  arrange(date, country)


  assign(paste("D2.6.SVY.LABR",i,sep="_"), temp)
}

D2.6.SVY.LABR<- bind_rows(D2.6.SVY.LABR_2016, D2.6.SVY.LABR_2017, D2.6.SVY.LABR_2018,D2.6.SVY.LABR_2019)

D2.6.SVY.LABR_wide <- D2.6.SVY.LABR %>%
  pivot_wider(names_from=date,
              names_sep = "_",
              values_from=c('LABR.SURVEYS', 'SPI.D2.6.LABR')
              ) %>%
  left_join(country_list)

write_excel_csv(D2.6.SVY.LABR_wide,
                path = paste(csv_output, "D2.6.SVY.LABR_indicator.csv", sep="/" ))
```




### 2.7 Health/Demographic survey

Health surveys collect information on various aspects of health of
populations, such as health expenditure, access, utilization, and outcomes.
They typically include Demographic and Health Surveys.  It is recommended that
health surveys be conducted at least every 3 to 5 years.

1 Point. 3 or more surveys done within past 10 years

0 Points. None within past 10 years

```{r svy.hlth}

#read in csv file.
D2.7.SVY.HLTH_df <- read_csv(file = paste(csv_dir, "D2.7.SVY.HLTH.csv", sep="/" )) %>%
  group_by(iso3c, country) %>%  # The next chunk of code will split our string with the years of the survey (i.e. "2000, 2010") in to separate rows.  We will then aggregate up.
  nest() %>%
  mutate(
    temp_col = map(
      data,
      ~ str_extract_all(.x$HLTH.SURVEYS, "\\d{4}") %>%
        flatten() %>%
        map_chr(~return(.x)) %>%
        as_tibble()
    )
  ) %>%
  unnest(keep_empty = TRUE) %>% # Now we have a database with the observations equal to Country*Census observations.  From here we can calculate latest census, etc.
  mutate(survey_date=as.numeric(value))

# Now calculate our SPI score for this indicator
for (i in 2016:2019) {
 temp <- D2.7.SVY.HLTH_df %>%
  mutate(date=i) %>%
  mutate(recency_indicator=((date-survey_date<=10) & (date-survey_date>0)) ) %>% #restrict to surveys inside 10 year period
  group_by(iso3c, country, date, HLTH.SURVEYS) %>% #group by country and create indicator for how many surveys over 10 year period
  summarise(SPI.D2.7.HLTH=case_when(
    sum(recency_indicator)>=3 ~ 1,
   # sum(recency_indicator)==2 ~ 0.6,
   # sum(recency_indicator)==1 ~ 0.3,
    TRUE ~ 0 )
  )  %>%
  select(iso3c, country, date, HLTH.SURVEYS, SPI.D2.7.HLTH  ) %>%
  arrange(date, country)


  assign(paste("D2.7.SVY.HLTH",i,sep="_"), temp)
}


D2.7.SVY.HLTH<- bind_rows(D2.7.SVY.HLTH_2016, D2.7.SVY.HLTH_2017, D2.7.SVY.HLTH_2018,D2.7.SVY.HLTH_2019)

D2.7.SVY.HLTH_wide <- D2.7.SVY.HLTH %>%
  pivot_wider(names_from=date,
              names_sep = "_",
              values_from=c('HLTH.SURVEYS', 'SPI.D2.7.HLTH')
              ) %>%
  left_join(country_list)

write_excel_csv(D2.7.SVY.HLTH_wide,
                path = paste(csv_output, "D2.7.SVY.HLTH_indicator.csv", sep="/" ))
```



### 2.8 Business/establishment survey

The business/establishment survey provides information on employment, hours,
and earnings of employees from a sample of business establishments including
private and public, entities that are classified based on an establishment's
principal activity from the business or establishment census.  Establishment
surveys include surveys of businesses, farms, and institutions.  They may ask
for information about the establishment itself and/or employee characteristics
and demographics.

1 Point. 3 or more surveys done within past 10 years

0 Points. None within past 10 years



```{r svy.bizz}

#read in csv file.
D2.8.SVY.BIZZ_df <- read_csv(file = paste(csv_dir, "D2.8.SVY.BIZZ.csv", sep="/" )) %>%
  group_by(iso3c, country) %>%  # The next chunk of code will split our string with the years of the survey (i.e. "2000, 2010") in to separate rows.  We will then aggregate up.
  nest() %>%
  mutate(
    temp_col = map(
      data,
      ~ str_extract_all(.x$BIZZ.SURVEYS, "\\d{4}") %>%
        flatten() %>%
        map_chr(~return(.x)) %>%
        as_tibble()
    )
  ) %>%
  unnest(keep_empty = TRUE) %>% # Now we have a database with the observations equal to Country*Census observations.  From here we can calculate latest census, etc.
  mutate(survey_date=as.numeric(value))

# Now calculate our SPI score for this indicator

for (i in 2016:2019) {
 temp <- D2.8.SVY.BIZZ_df %>%
  mutate(date=i) %>%
  mutate(recency_indicator=((date-survey_date<=10) & (date-survey_date>0)) ) %>% #restrict to surveys inside 10 year period
  group_by(iso3c, country, date, BIZZ.SURVEYS) %>% #group by country and create indicator for how many surveys over 10 year period
  summarise(SPI.D2.8.BIZZ=case_when(
    sum(recency_indicator)>=3 ~ 1,
    #sum(recency_indicator)==2 ~ 0.6,
    #sum(recency_indicator)==1 ~ 0.3,
    TRUE ~ 0 )
  )  %>%
  select(iso3c, country, date, BIZZ.SURVEYS, SPI.D2.8.BIZZ  ) %>%
  arrange(date, country)


  assign(paste("D2.8.SVY.BIZZ",i,sep="_"), temp)
}


D2.8.SVY.BIZZ<- bind_rows(D2.8.SVY.BIZZ_2016, D2.8.SVY.BIZZ_2017, D2.8.SVY.BIZZ_2018,D2.8.SVY.BIZZ_2019)

D2.8.SVY.BIZZ_wide <- D2.8.SVY.BIZZ %>%
  pivot_wider(names_from=date,
              names_sep = "_",
              values_from=c('BIZZ.SURVEYS', 'SPI.D2.8.BIZZ')
              ) %>%
  left_join(country_list)

write_excel_csv(D2.8.SVY.BIZZ_wide,
                path = paste(csv_output, "D2.8.SVY.BIZZ_indicator.csv", sep="/" ))
```



### CS Score

CS Country  Score=Weighted Score÷Maximum Category Score×100

```{r cs}
# join datasets together based on country and date
D2.CS <- D2.1.CEN.POPU %>%
  left_join(D2.2.CEN.AGRI) %>%
  left_join(D2.3.CEN.BIZZ) %>%
  left_join(D2.4.SVY.HOUS) %>%
  left_join(D2.5.SVY.AGRI) %>%
  left_join(D2.6.SVY.LABR) %>%
  left_join(D2.7.SVY.HLTH) %>%
  left_join(D2.8.SVY.BIZZ) %>%
  select(iso3c, country, date, starts_with("SPI.D2"))

#Now calculate MSC score which is the average across the 12 indicators
D2.CS <- D2.CS %>%
  mutate(SPI.D2.CS=100*rowMeans(.[grep(x=colnames(D2.CS),
                                              pattern="SPI.D2")], na.rm=TRUE)) %>%
  arrange(date, iso3c) %>%
  select(iso3c, country, date,SPI.D2.CS, starts_with("SPI.D2"))


D2.CS_wide <- D2.CS %>%
  select(iso3c, country, date,SPI.D2.CS) %>%
  pivot_wider(names_from=date,
              names_prefix = "SPI.D2.CS_",
              values_from=c('SPI.D2.CS')
              )

write_excel_csv(D2.CS_wide,
                path = paste(csv_output, "D2.CS_indicator.csv", sep="/" ))

```

Produce summary statistics of our indicators.

```{r}

msc_sumstats <- D2.CS %>%
  group_by(date) %>%
  select(SPI.D2.CS, starts_with("SPI.D2")) %>%
  skim() %>%
  select(c("skim_variable",
           "date", "complete_rate", "numeric.mean",  "numeric.sd",    "numeric.p0",    "numeric.p25",   "numeric.p50",
           "numeric.p75",   "numeric.p100",  "numeric.hist"))

kable(msc_sumstats, caption="Summary Statistics of Census and Surveys Indicator Variables" , col.names = c("Indicator", "Date", "Completion Rate", "Mean",  "Std Dev","Min", "25th Percentile", "Median", "75th Percentile", "Max",  "Histogram"), digits = 2 ) %>%
  kable_styling(bootstrap_options = c("striped", "hover"))

```



## SPI Dimension 3: Availability of Key Indicators (AKI):

Below is the list of AKI indicators:

- AKI 3.1: Poverty headcount ratio at $1.90 a day (2011 PPP) (% of population)  
- AKI 3.2: Food Insecurity Experience Scale 
- AKI 3.3: Mortality rate, under-5 (per 1,000 live births)  
- AKI 3.4: Proportion of children and young people in grades 2 or 3 achieving at least a minimum proficiency level in reading and mathematics, by sex.  
- AKI 3.5: Maternal Mortality
- AKI 3.6: People using safely managed drinking water services (% of population)  
- AKI 3.7: Access to electricity (% of population)  
- AKI 3.8: Unemployment, total (% of total labor force)  
- AKI 3.9: Manufacturing, value added (% of GDP)  
- AKI 3.10: Annualized average growth rate in per capita real survey mean consumption or income, bottom 40% of population (%)  
- AKI 3.11: Level of water stress: freshwater withdrawal as a proportion of available freshwater resources  
- AKI 3.12: Renewable energy consumption (% of total final energy consumption)  
- AKI 3.13: Households and NPISHs Final consumption expenditure (current LCU)  
- AKI 3.14: Quarterly GDP
- AKI 3.15: Debt service (PPG and IMF only, % of exports of goods, services and primary income) 

```{r metadata_pull}

# pull in some WDI metadata which will be used for constructing AKI indicators.
# WDI metadata was gathered using previous vintages of WDI from 2016-19
# Metadata was gathered and saved as csv files in the 011_data folder

for (i in 2016:2019) {
  temp <- read_csv(file = paste(csv_dir, "/WDI_metadata_",i,".csv", sep="" )) %>%
    as_tibble(.name_repair='universal') %>%
    mutate(National.accounts.reference.year=as.numeric(National.accounts.reference.year),
           date=i)
  
    assign(paste("WDI_metadata",i,sep="_"), temp)
}

WDI_metadata <- bind_rows(WDI_metadata_2016, WDI_metadata_2017, WDI_metadata_2018, WDI_metadata_2019)
```



### AKI 3.1, 3.9, 3.10, 3.11, 3.12, 3.13

First we will pull data for indicators coming straight from WDI.  For some indicators, we will use alternative sources.

Indicators coming from the WDI directly are:  

- AKI 3.1: Poverty headcount ratio at $1.90 a day (2011 PPP) (% of population)  
- AKI 3.9: Manufacturing, value added (% of GDP)  
- AKI 3.10: Annualized average growth rate in per capita real survey mean consumption or income, bottom 40% of population (%)  
- AKI 3.11: Level of water stress: freshwater withdrawal as a proportion of available freshwater resources  
- AKI 3.12: Renewable energy consumption (% of total final energy consumption)  
- AKI 3.13: Households and NPISHs Final consumption expenditure (current LCU)  

Scoring:

1 Point. 3 or more values available within past 5 years	

0 Points. None within past 5 years

```{r data_pulls, echo=TRUE}

##########
# Pull Tags for WDI Data
##########

# make request to World Bank API
wdiRequest <- GET(url = "http://api.worldbank.org/v2/indicator?per_page=20000&format=json&source=2")
wdiResponse <- content(wdiRequest, as = "text", encoding = "UTF-8")

# Parse the JSON content and convert it to a data frame.
wdisJSON <- jsonlite::fromJSON(wdiResponse, flatten = TRUE) %>%
    data.frame()

EdStatsRequest <- GET(url = "http://api.worldbank.org/v2/indicator?per_page=20000&format=json&source=12")
EdStatsResponse <- content(EdStatsRequest, as = "text", encoding = "UTF-8")

# Parse the JSON content and convert it to a data frame.
EdStatsJSON <- jsonlite::fromJSON(EdStatsResponse, flatten = TRUE) %>%
  data.frame()

aki<- c(
        'Pupils below minimum reading proficiency at end of primary (%). Low GAML threshold',
        'People using safely managed drinking water services (% of population)',
        'Unemployment, total (% of total labor force) (national estimate)' ,
        'Manufacturing, value added (% of GDP)',
        'Annualized average growth rate in per capita real survey mean consumption or income, bottom 40% of population (%)',
        'Level of water stress: freshwater withdrawal as a proportion of available freshwater resources',
        'Renewable energy consumption (% of total final energy consumption)',
        'Households and NPISHs Final consumption expenditure (current LCU)',
        'Debt service (PPG and IMF only, % of exports of goods, services and primary income)'  )

get_tag_aki_df<-wdisJSON %>%
  bind_rows(EdStatsJSON) %>%
  filter((name %in% aki	)) %>%
  arrange(factor(name, levels = aki)) %>%
  select(id, name,  sourceOrganization) 

#get WDI metadata infor
cache_list<-wbstats::wbcache()

aki_list<-get_tag_aki_df[,'id']

for (reference_year in 2016:2019) {
  temp <-wbstats::wb(country="countries_only", 
              indicator=aki_list,
              startdate=reference_year-5,
              enddate=reference_year,
              return_wide = T,
              removeNA=FALSE) %>%
          filter(((reference_year-as.numeric(date))<=5) & (reference_year>=as.numeric(date))) %>% #filter out years outside reference window of 3 years     
          write_excel_csv(path = paste(csv_output, "/D3.AKI_data_pull_",reference_year,".csv", sep="" )) %>%
          mutate_at(.vars=aki_list, ~if_else(is.na(.),0,1)) %>% #create 0,1 variable for whether data point exists for country
          group_by(iso3c, country) %>%
          summarise_all((~(if(is.numeric(.)) sum(., na.rm = TRUE) else first(.)))) %>% #group by country to create one observation per country                 containing whether or not data point existed
          mutate_at(.vars=aki_list, ~case_when(
            .>=3 ~ 1,
            #.==2 ~ 0.6,
            #.==1 ~ 0.3,
            .==0 ~ 0, 
            TRUE ~ 0
          )) %>% # 1 point for at least 3 values, 0.6 for 2 values, 0.3 for 1 values, 0 otherwise
          mutate(date=reference_year) %>%
          ungroup() %>%
          select(iso3c, country, date,  aki_list) %>%
          rename_at(aki_list, ~(paste('SPI.D3.',.,sep=""))) %>% #add 'D3.' as prefix before these indicators.
          left_join(country_list) #attach country metadata


  assign(paste("D3.AKI",reference_year,sep="_"), temp)
}



```

### AKI 3.1: Poverty headcount ratio at $1.90 a day (2011 PPP) (% of population)

The data will be pulled from the WDI and combined with metadata from the Povcalnet

Scoring is as follows.  Overall score is the average of quality and frequency:
Quality (0.5 points total):
1 Point. Comparable data lasting at least two years within past 5 years	
0 Points. No comparable data within past 5 years

Frequency (0.5 points total):
1 Point. 3 or more values available within past 5 years	
0 Points. None within past 5 years


```{r 3.1}
library(povcalnetR) 

#the code below is based on public released code by povcalnet team:
#https://github.com/worldbank/Global_Poverty_Blogs/blob/master/bg_povcalnet_comparability/R/comparability_breaks.R

# some contants
year_range <- 1990:2020
metadata_path <- "https://development-data-hub-s3-public.s3.amazonaws.com/ddhfiles/506801/povcalnet_comparability.csv"

# Load data ---------------------------------------------------------------

metadata <- read_csv(metadata_path)
cov_lkup <- c(3, 2, 1, 4)
names(cov_lkup) <- c("N", "U", "R", "A")

dat_lkup <- c(2,1)
names(dat_lkup) <- c("income","consumption")


pcn <- povcalnet()
pcn$coveragetype <- cov_lkup[pcn$coveragetype]
pcn$datatype <- dat_lkup[pcn$datatype]

df <- pcn %>%
  inner_join(metadata, by = c("countrycode", "year", "coveragetype", "datatype"))


#Now loop from 2016 and 2019, keeping just data inside last 5 years.
for (reference_year in 2016:2019) {
  
  temp1<-df %>%
        filter(coveragetype %in% c(3,4)) %>% #keep just nationally representative samples
        mutate(frequency=((reference_year-as.numeric(year))<=5) & (reference_year>=as.numeric(year))) %>% 
        filter(frequency==TRUE) %>%
        group_by(countrycode, comparability) %>% #for each country and comparability type, get number of comparable estimates
        summarise(SPI.QUAL.D3.POV=n())  %>%
        ungroup() %>%
        group_by(countrycode) %>% #now get a total by country with the max number of comparable estimates
        summarise(SPI.QUAL.D3.POV=max(SPI.QUAL.D3.POV, na.rm=T)) %>%
        mutate(SPI.QUAL.D3.POV=if_else(SPI.QUAL.D3.POV>=2,1,0)) #only give point if there is at least two observations that are comparable
  
  
  temp <-df %>%
          filter(coveragetype %in% c(3,4)) %>% #keep just nationally representative samples
          mutate(frequency=((reference_year-as.numeric(year))<=5) & (reference_year>=as.numeric(year))) %>% 
          mutate(SPI.FREQ.D3.POV=if_else(frequency==TRUE,1,0)) %>% #create 0,1 variable for whether data point exists for country
          group_by(countryname,countrycode) %>%
          summarise(SPI.FREQ.D3.POV=sum(SPI.FREQ.D3.POV, na.rm=T)) %>% 
          mutate(SPI.FREQ.D3.POV=case_when(
            SPI.FREQ.D3.POV>=3 ~ 1,
           # SPI.FREQ.D3.POV==2 ~ 0.3,
           # SPI.FREQ.D3.POV==1 ~ 0.15,
            SPI.FREQ.D3.POV==0 ~ 0, 
            TRUE ~ 0
          )) %>%
          left_join(temp1) %>% #now bring in quality dimension
          mutate(SPI.QUAL.D3.POV=if_else(!is.na(SPI.QUAL.D3.POV),SPI.QUAL.D3.POV,0)) %>% #recode some missing values as zero, who had no observations in 5 year window
          mutate(SPI.D3.POV=(SPI.FREQ.D3.POV+SPI.QUAL.D3.POV)/2) %>% 
          mutate(date=reference_year,
                 country=countryname,
                 iso3c=countrycode) %>%
          select( country,iso3c, date, starts_with('SPI.D3.POV')) 


  assign(paste("D3.1.AKI",reference_year,sep="_"), temp)
}

D3.1.AKI <- bind_rows(D3.1.AKI_2016, D3.1.AKI_2017, D3.1.AKI_2018, D3.1.AKI_2019)



```

### AKI 3.2: Hunger

We examine the Food Insecurity Experience Scale from the FAO (http://www.fao.org/faostat/en/#data/FS).  Data for food insecurity was pulled on May 1, 2020.

Scoring
1 Point. 3 or more values available within past 5 years	
0 Points. None within past 5 years


```{r AKI_3.2}


#Read in data from UN Inter-agency Group for Child Mortality Estimation

D3.2.AKI.FIES <- read_csv(file=paste(csv_dir, 'D3.2.FIES.csv', sep="/")) %>%
  rename(year_code="Year Code",
         iso3c="ISO3 Code") %>%
  mutate(date=as.numeric(str_extract(year_code, "^.{4}")),
         SPI.D3.FIES=Value)

#Now loop from 2016 and 2019, keeping just data inside last 5 years.
for (reference_year in 2016:2019) {
  temp <-D3.2.AKI.FIES %>%
          filter(((reference_year-as.numeric(date))<=5) & (reference_year>=as.numeric(date))) %>% #filter out years outside reference window of 5 years     
          mutate(SPI.D3.FIES=if_else(is.na(SPI.D3.FIES),0,1)) %>% #create 0,1 variable for whether data point exists for country
          group_by(iso3c) %>%
          summarise_all((~(if(is.numeric(.)) sum(., na.rm = TRUE) else first(.)))) %>% #group by country to create one observation per country                 containing whether or not data point existed
          mutate(SPI.D3.FIES=case_when(
            SPI.D3.FIES>=3 ~ 1,
           # SPI.D3.FIES==2 ~ 0.6,
           # SPI.D3.FIES==1 ~ 0.3,
            SPI.D3.FIES==0 ~ 0, 
            TRUE ~ 0
          )) %>% # 1 point for at least 3 values, 0.6 for 2 values, 0.3 for 1 values, 0 otherwise
          mutate(date=reference_year) %>%
          select( iso3c, date, starts_with('SPI.')) 


  assign(paste("D3.2.AKI",reference_year,sep="_"), temp)
}

D3.2.AKI <- bind_rows(D3.2.AKI_2016, D3.2.AKI_2017, D3.2.AKI_2018, D3.2.AKI_2019)
```



### AKI 3.3: Mortality rate, under-5 (per 1,000 live births)  

The data in WDI is modeled, based on HHS and Vital Registration.  We use the following as a source for raw data produced by national statistical offices that is used in this modeling.  

https://childmortality.org/data 

Data was pulled on April 13, 2020.

Countries are ranked by source quality (admin data > survey > no data) and frequency.
Score on this indicator has a max of 1 pointand is the average of the quality and frequency indicators.  Detailed scoring for source quality and frequency components are below.

Quality (0.5 points total):
1 Point. Vital Registration data available within past 5 years	
0 Points. None within past 5 years

Frequency (0.5 points total):
1 Point. 3 or more values available within past 5 years	
0 Points. None within past 5 years


```{r aki_3.3}

#Read in data from UN Inter-agency Group for Child Mortality Estimation

D3.3.AKI.MORT <- read_csv(file=paste(csv_dir, 'D3.3_child_mort_RAW_v01.csv', sep="/")) %>%
  filter(INDICATOR=='Under-five mortality rate' & SEX=='Total') %>% #keep just observations for under 5 child mortality and for both sexes 
  filter(OBS_STATUS=='Included in IGME') %>% # Also keep only surveys that met IGME criteria for inclusion as a nationally representative statistic
  mutate(date=as.numeric(str_extract(TIME_PERIOD, "^.{4}")),
         country=REF_AREA,
         D3.CHLD.MORT=OBS_VALUE)

#Now loop from 2016 and 2019, keeping just data inside last 5 years.
for (reference_year in 2016:2019) {
  temp <-D3.3.AKI.MORT %>%
          mutate(frequency=((reference_year-as.numeric(date))<=5) & (reference_year>=as.numeric(date))) %>% 
          mutate(SPI.FREQ.D3.CHLD.MORT=if_else(frequency==TRUE,1,0)) %>% #create 0,1 variable for whether data point exists for country
          mutate(SPI.QUAL.D3.CHLD.MORT=case_when(
                   reference_year-as.numeric(date)>=5 ~ 0, #outside 5 year window
                   SERIES_METHOD=='Vital Registration' ~ 1,
                  # SERIES_METHOD!='Vital Registration' ~ 0.25,
                   TRUE ~ 0
                 )) %>%
          group_by(country) %>%
          summarise(SPI.FREQ.D3.CHLD.MORT=sum(SPI.FREQ.D3.CHLD.MORT, na.rm=T),
                    SPI.QUAL.D3.CHLD.MORT=max(SPI.QUAL.D3.CHLD.MORT, na.rm=T)) %>% 
          mutate(SPI.FREQ.D3.CHLD.MORT=case_when(
            SPI.FREQ.D3.CHLD.MORT>=3 ~ 1,
           # SPI.FREQ.D3.CHLD.MORT==2 ~ 0.3,
            #SPI.FREQ.D3.CHLD.MORT==1 ~ 0.15,
            SPI.FREQ.D3.CHLD.MORT==0 ~ 0, 
            TRUE ~ 0
          )) %>%
          mutate(SPI.D3.CHLD.MORT=(SPI.FREQ.D3.CHLD.MORT+SPI.QUAL.D3.CHLD.MORT)/2) %>% 
          mutate(date=reference_year) %>%
          select( country, date, starts_with('SPI.')) 


  assign(paste("D3.3.AKI",reference_year,sep="_"), temp)
}

D3.3.AKI <- bind_rows(D3.3.AKI_2016, D3.3.AKI_2017, D3.3.AKI_2018, D3.3.AKI_2019)


```


```{r country_label_fixes, include=FALSE}

#make country labels comparable to WDI
D3.3.AKI <- D3.3.AKI %>%
  mutate(country=case_when(
    country=="Bahamas" ~ "Bahamas, The",                  
    country=="Bolivia (Plurinational State of)"   ~  "Bolivia"   ,                  
    country=="Côte d'Ivoire"  ~  "Cote d'Ivoire"                ,
    country=="Democratic Republic of the Congo"  ~ "Congo, Dem. Rep."  ,           
    country=="Congo"  ~ "Congo, Rep."        ,           
    country=="Curacao"  ~  "Curacao"       ,                
    country=="Czechia"   ~  "Czech Republic"     ,        
    country=="Egypt"  ~ "Egypt, Arab Rep."        ,      
    country=="Micronesia (Federated States of)"  ~ "Micronesia, Fed. Sts."  ,       
    country=="United Kingdom"   ~  "United Kingdom"    ,         
    country=="Gambia"    ~  "Gambia, The"     ,           
    country=="Iran (Islamic Republic of)"  ~ "Iran, Islamic Rep." ,           
    country=="Kyrgyzstan"   ~  "Kyrgyz Republic"   ,          
    country=="Republic of Korea"   ~ "Korea, Rep."   ,               
    country=="Lao People's Democratic Republic"    ~  "Lao PDR" ,                    
    country=="Saint Kitts and Nevis	"   ~ "St. Kitts and Nevis",
    country=="Saint Lucia"   ~ "St. Lucia",
    country=="Republic of Moldova"   ~  "Moldova"   ,                  
    country=="Democratic People's Republic of Korea"  ~  "Korea, Dem. People’s Rep." ,  
    country=="Slovakia"   ~  "Slovak Republic"             ,
    country=="United Republic of Tanzania"   ~  "Tanzania" ,                    
    country=="United States of America"  ~ "United States"  ,                
    country=="Saint Vincent and the Grenadines" ~ "St. Vincent and the Grenadines" ,
    country=="Venezuela (Bolivarian Republic of)"  ~  "Venezuela, RB"   ,              
    country=="British Virgin Islands"  ~  "British Virgin Islands" ,       
    country=="Viet Nam"  ~ "Vietnam"        ,              
    country=="Yemen"  ~ "Yemen, Rep.",
    country=="United Kingdom of Great Britain and Northern Ireland" ~ "United Kingdom",
    TRUE ~ country
  ))


```




```{r AKI_3.3, eval=FALSE, include=FALSE}
EdStatsRequest_metadata <- GET(url = "https://api.worldbank.org/v2/country/all/indicator/SE.LPV.PRIM.BMP?footnote=y&per_page=30000&format=json")
EdStatsResponse_metadata <- content(EdStatsRequest_metadata, as = "text", encoding = "UTF-8")

# Parse the JSON content and convert it to a data frame.
EdStatsJSON_metadata <- jsonlite::fromJSON(EdStatsResponse_metadata, flatten = TRUE) %>%
  data.frame() %>%
  filter(str_length(footnote)>0)
```


### AKI 3.5 Maternal Mortality Ratio

The data is sourced from the the Inter-Agency Group (200 countries) to improve time and country coverages. Tne national estimates with the source information are available at https://www.who.int/reproductivehealth/publications/maternal-mortality-2000-2017/en/.  The data was accessed on April 3, 2020.  This data was used for modelling maternal mortality, which are then fed into the modelled estimates on Maternal Mortality (SH.STA.MMRT) in the WDI.

For our purposes, we will use the raw data from the country surveys as our data for the availability of key indicators, since the modelled estimates are produced out of sample, and thus not reflective of a country's national statistical system.

Countries are ranked by source quality (admin data > survey > no data) and frequency.
Score on this indicator has a max of 1 point and is the average of the frequency and quality indicators.  Detailed scoring for source quality and frequency components are below.

Quality (1 points total):
1 Point. Vital Registration data available within past 5 years	
0 Points. None within past 5 years

Frequency (1 points total):
1 Point. 3 or more values available within past 5 years	
0 Points. None within past 5 years

```{r AKI_3.5}


#Read in data from UN Inter-agency Group 

D3.5.AKI.MMR <- read_csv(file=paste(csv_dir, 'D3.5.BMat2019_datainputs.csv', sep="/")) %>%
  filter(modelinclude=='TRUE') %>% #keep just observations that the UN inter-agency group deemed credible for use in modelling 
  mutate(date=end,
         iso3c=iso,
         D3.MMRT=pm_obs)

#Now loop from 2016 and 2019, keeping just data inside last 5 years.
for (reference_year in 2016:2019) {
  temp <-D3.5.AKI.MMR %>%
          mutate(frequency=((reference_year-as.numeric(date))<=5) & (reference_year>=as.numeric(date))) %>% 
          mutate(SPI.FREQ.D3.MMRT=if_else(frequency==TRUE,1,0)) %>% #create 0,1 variable for whether data point exists for country
          mutate(SPI.QUAL.D3.MMRT=case_when(
                   reference_year-as.numeric(date)>=5 ~ 0, #outside 5 year window
                   type=='vr' ~ 1,
                   #type!='vr' ~ 0.25,
                   TRUE ~ 0
                 )) %>%
          group_by(iso3c) %>%
          summarise(SPI.FREQ.D3.MMRT=sum(SPI.FREQ.D3.MMRT, na.rm=T),
                    SPI.QUAL.D3.MMRT=max(SPI.QUAL.D3.MMRT, na.rm=T)) %>% 
          mutate(SPI.FREQ.D3.MMRT=case_when(
            SPI.FREQ.D3.MMRT>=3 ~ 1,
  #          SPI.FREQ.D3.MMRT==2 ~ 0.3,
  #          SPI.FREQ.D3.MMRT==1 ~ 0.15,
            SPI.FREQ.D3.MMRT==0 ~ 0, 
            TRUE ~ 0
          )) %>%
          mutate(SPI.D3.MMRT=(SPI.FREQ.D3.MMRT+SPI.QUAL.D3.MMRT)/2) %>% 
          mutate(date=reference_year) %>%
          select( iso3c, date, starts_with('SPI.')) 


  assign(paste("D3.5.AKI",reference_year,sep="_"), temp)
}

D3.5.AKI <- bind_rows(D3.5.AKI_2016, D3.5.AKI_2017, D3.5.AKI_2018, D3.5.AKI_2019)

```






### AKI 3.7: Access to electricity (% of population) 

Access to Electricity (% of population) has 99% coverage in the WDI database making it less useful for distinguishing between countries.  In order to improve the usefulness of this indicator, we are making use of data compiled by World Bank colleagues on surveys containing electricity items.  According to the access to electricity metadata, around 42 countries have no data and an imputed value is assigned based on the regional average.  By going to survey source data, we address this issue of imputation, which gives a misleading picture of availability for our purposes.  Below is the metadata from the WDI for methods for the access to electricity indicator.

>Data for access to electricity are collected among different sources: mostly data from nationally representative household surveys (including national censuses) were used. Survey sources include Demographic and Health Surveys (DHS) and Living Standards Measurement Surveys (LSMS), Multi-Indicator Cluster Surveys (MICS), the World Health Survey (WHS), other nationally developed and implemented surveys, and various government agencies (for example, ministries of energy and utilities). Given the low frequency and the regional distribution of some surveys, a number of countries have gaps in available data. To develop the historical evolution and starting point of electrification rates, a simple modeling approach was adopted to fill in the missing data points - around 1990, around 2000, and around 2010. Therefore, a country can have a continuum of zero to three data points. There are 42 countries with zero data point and the weighted regional average was used as an estimate for electrification in each of the data periods. 170 countries have between one and three data points and missing data are estimated by using a model with region, country, and time variables. The model keeps the original observation if data is available for any of the time periods. This modeling approach allowed the estimation of electrification rates for 212 countries over these three time periods (Indicated as "Estimate"). Notation "Assumption" refers to the assumption of universal access in countries classified as developed by the United Nations. Data begins from the year in which the first survey data is available for each country.

Scoring
1 Point. 3 or more values available within past 5 years	
0 Points. None within past 5 years

```{r D3.7_AKI}
#read in file from rawdata folder
D3.7.ELEC <- read_csv(file = paste(csv_dir, "D3.7.ELEC.csv", sep="/" )) %>%
  rename(country=Country,
         iso3c="Country code",
         date=Time)

#Now loop from 2016 and 2019, keeping just data inside last 5 years.
for (reference_year in 2016:2019) {
  temp <-D3.7.ELEC %>%
          mutate(frequency=((reference_year-as.numeric(date))<=5) & (reference_year>=as.numeric(date))) %>% 
          mutate(SPI.D3.ELEC=if_else(frequency==TRUE,1,0)) %>% #create 0,1 variable for whether data point exists for country
          group_by(country) %>%
          summarise(SPI.D3.ELEC=sum(SPI.D3.ELEC, na.rm=T)) %>%
          mutate(SPI.D3.ELEC=case_when(
            SPI.D3.ELEC>=3 ~ 1,
 #           SPI.D3.ELEC==2 ~ 0.6,
 #           SPI.D3.ELEC==1 ~ 0.3,
            SPI.D3.ELEC==0 ~ 0, 
            TRUE ~ 0
          )) %>%
          mutate(date=reference_year) %>%
          select( country, date, starts_with('SPI.')) 


  assign(paste("D3.7.AKI",reference_year,sep="_"), temp)
}

D3.7.AKI <- bind_rows(D3.7.AKI_2016, D3.7.AKI_2017, D3.7.AKI_2018, D3.7.AKI_2019)
```



### AKI 3.14 Quarterly GDP

Quarterly GDP numbers were pulled from the IMF website.

After GDP by expenditure, quarterly GDP is probably the most important development to be made in a system of National Accounts, before the development of full sectoral accounts. IMF IFS has quarterly GDP from various sources, including governments and international agencies at https://data.imf.org/?sk=4C514D48-B6BA-49ED-8AB9-52B0C1A0179B

Scoring
1 Point. 3 or more values available within past 5 years	
0 Points. None within past 5 years



```{r AKI_3.17}

#read in file from rawdata folder
D3.14.QUART.GDP <- read_csv(file = paste(csv_dir, "D3.14.QUART.GDP.csv", sep="/" ))

#clean data and produce indicator for each year
#Now loop from 2016 and 2019, keeping just data inside last 5 years.
for (reference_year in 2016:2019) {
temp <-D3.14.QUART.GDP %>%
  mutate(date=as.numeric(str_sub(TIME_PERIOD,1,4))) %>%
  mutate(iso2c=REF_AREA) %>%
  mutate(frequency=((reference_year-as.numeric(date))<=5) & (reference_year>=as.numeric(date))) %>% 
  mutate(SPI.D3.QUART.GDP=if_else(frequency==TRUE,1,0)) %>% #create 0,1 variable for whether data point exists for country
  group_by(iso2c, date) %>%
  summarise(SPI.D3.QUART.GDP=max(SPI.D3.QUART.GDP, na.rm=T)) %>% 
  group_by(iso2c) %>%
  summarise(SPI.D3.QUART.GDP=sum(SPI.D3.QUART.GDP, na.rm=T)) %>% 
  mutate(SPI.D3.QUART.GDP=case_when(
    SPI.D3.QUART.GDP>=3 ~ 1,
    #SPI.D3.QUART.GDP==2 ~ 0.6,
    #SPI.D3.QUART.GDP==1 ~ 0.3,
    SPI.D3.QUART.GDP==0 ~ 0, 
    TRUE ~ 0
  )) %>%
  mutate(date=reference_year) %>%
  select( iso2c, date, starts_with('SPI.D3.QUART.GDP')) 


  assign(paste("D3.14.AKI",reference_year,sep="_"), temp)
}

D3.17.AKI <- bind_rows(D3.14.AKI_2016, D3.14.AKI_2017, D3.14.AKI_2018, D3.14.AKI_2019)  

```

### AKI 3.15 Debt service (PPG and IMF only, % of exports of goods, services and primary income)

For this indicator, Debt service (PPG and IMF only, % of exports of goods, services and primary income), we will pull data from the WDI but modify the scoring using the WDI metadata on whether the external debt data is actual, estimated, or preliminary.  The status “as reported (actual)” indicates that the country was fully current in its reporting under the DRS and that World Bank staff are satisfied that the reported data give an adequate and fair representation of the country’s total public debt. “Preliminary” data are based on reported or collected information, but because of incompleteness or other reasons, an element of staff estimation is included. “Estimated” data indicate that countries are not current in their reporting and that a significant element of staff estimation has been necessary for producing the data tables. 

Scoring is as follows with the overall score being the average of the frequency and quality dimensions:

Frequency:

1 Point. 3 or more values available within past 5 years	
0 Points. None within past 5 years

Quality: 

1 Points. Actual value
0 Points. No value

```{r 3.15_debt}

#reshape metaadata file
metadata_3.15 <- WDI_metadata %>%
  transmute(iso3c=if_else(is.na(Country.Code), Code, Country.Code),
            date=date,
            External_debt_Reporting=External.debt.Reporting.status) %>%
  mutate(SPI.QUAL.D3.DT.TDS.DPPF.XP.ZS= case_when(
            External_debt_Reporting=='Actual' ~ 1,
           # External_debt_Reporting=='Preliminary' ~ 0.3,
            #External_debt_Reporting=='Estimate' ~ 0.15,
            TRUE ~ 0
          ))

#pull data from wdi and merge with metadata
for (reference_year in 2016:2019) {
  temp <-wbstats::wb(country="countries_only", 
              indicator='DT.TDS.DPPF.XP.ZS',
              startdate=reference_year-5,
              enddate=reference_year,
              return_wide = T,
              removeNA=FALSE) %>%
          filter(((reference_year-as.numeric(date))<=5) & (reference_year>=as.numeric(date))) %>% #filter out years outside reference window of 3 years     
          mutate_at(.vars=c('DT.TDS.DPPF.XP.ZS'), ~if_else(is.na(.),0,1)) %>% #create 0,1 variable for whether data point exists for country
          group_by(iso3c, country) %>%
          summarise_all((~(if(is.numeric(.)) sum(., na.rm = TRUE) else first(.)))) %>% #group by country to create one observation per country                 containing whether or not data point existed
          mutate(SPI.FREQ.D3.DT.TDS.DPPF.XP.ZS = case_when(
            DT.TDS.DPPF.XP.ZS>=3 ~ 1,
           # DT.TDS.DPPF.XP.ZS==2 ~ 0.3,
           # DT.TDS.DPPF.XP.ZS==1 ~ 0.15,
            DT.TDS.DPPF.XP.ZS==0 ~ 0, 
            TRUE ~ 0
          )) %>% # 0.5 point for at least 3 values, 0.3 for 2 values, 0.15 for 1 values, 0 otherwise
          mutate(date=reference_year) %>%
          left_join(metadata_3.15) %>% #attach country metadata 
          mutate(SPI.QUAL.D3.DT.TDS.DPPF.XP.ZS= case_when(
            External_debt_Reporting=='Actual' ~ 1,
           # External_debt_Reporting=='Preliminary' ~ 0.3,
            #External_debt_Reporting=='Estimate' ~ 0.15,
            TRUE ~ 0
          )) %>% # 0.5 point for actual, 0.3 for preliminary, 0.15 for estimate
          mutate(
                 SPI.D3.DT.TDS.DPPF.XP.ZS=(SPI.QUAL.D3.DT.TDS.DPPF.XP.ZS + SPI.FREQ.D3.DT.TDS.DPPF.XP.ZS)/2) %>%
          ungroup() %>%
          select(iso3c, country, date,  contains('D3.DT.TDS.DPPF.XP.ZS')) 



  assign(paste("D3.15.AKI",reference_year,sep="_"), temp)
}

D3.15.AKI <- bind_rows(D3.15.AKI_2016, D3.15.AKI_2017, D3.15.AKI_2018, D3.15.AKI_2019)  

```


### Combine all sources into AKI database

```{r aki_combine_write}

#now combine AKI databases and write to csv
D3.AKI <- bind_rows(D3.AKI_2016, D3.AKI_2017, D3.AKI_2018, D3.AKI_2019) %>%
  left_join(D3.1.AKI) %>%
  left_join(D3.2.AKI) %>%
  left_join(D3.3.AKI) %>%
  left_join(D3.5.AKI) %>%
  left_join(D3.7.AKI) %>%
  left_join(D3.15.AKI) %>%
  left_join(D3.17.AKI) %>%
  mutate_at(vars(starts_with('SPI.D3.')), ~if_else(is.na(.),0,as.numeric(.))) 

#calculate overall AKI score  
D3.AKI <- D3.AKI  %>%
  mutate(SPI.D3.AKI=100*rowSums(.[grep('SPI.D3.', colnames(D3.AKI))], na.rm=T)/length(grep('SPI.D3.', colnames(D3.AKI)))) 




D3.AKI_wide <- D3.AKI %>%
  select(iso3c, country, date,SPI.D3.AKI) %>%
  pivot_wider(names_from=date,
              names_prefix = "SPI.D3.AKI_",
              values_from=c('SPI.D3.AKI')
              ) %>%
  left_join(country_list)

write_excel_csv(D3.AKI_wide,
                path = paste(csv_output, "D3.AKI_indicator.csv", sep="/" ))

```


Produce summary statistics of our indicators.

```{r}

aki_sumstats <- D3.AKI %>%
  group_by(date) %>%
  select(starts_with('SPI.D3')) %>%
  skim() %>%
  select(c("skim_variable",
           "date", "complete_rate", "numeric.mean",  "numeric.sd",    "numeric.p0",    "numeric.p25",   "numeric.p50",
           "numeric.p75",   "numeric.p100",  "numeric.hist"))

kable(aki_sumstats, caption="Summary Statistics of Availability of Key Indicator Variables" , col.names = c("Indicator", "Date", "Completion Rate", "Mean",  "Std Dev","Min", "25th Percentile", "Median", "75th Percentile", "Max",  "Histogram"), digits = 2 ) %>%
  kable_styling(bootstrap_options = c("striped", "hover"))

```


## SPI Dimension 4: Dissemination Practices and Openness (DPO):


### 4.1 NSO has an Advance Release Calendar and it is published 

This indicator refers to a dissemination practice for relevant statistical
data in accordance to the published advance release calendar.  Its presence
provides information on upcoming releases in advance and creates public
awareness; this will lead to more discipline and accountability from the
statistical office.

1 Point. Yes
0 Points. No

```{r cald}

#read in csv file.
D4.1.SC.DPO.CALD <- read_csv(file = paste(csv_dir, "D4.1.SC.DPO.CALD.csv", sep="/" )) %>%
  mutate(SPI.D4.1.CALD=case_when(
    CALD==1 ~ 1, 
    CALD==0 ~ 0, 
    TRUE ~ 0 
  ))  %>%
  select(iso3c, country, date, CALD, SPI.D4.1.CALD  ) %>%
  arrange(date, country)

D4.1.SC.DPO.CALD_wide <- D4.1.SC.DPO.CALD %>%
  pivot_wider(names_from=date,
              names_sep = "_",
              values_from=c('CALD', 'SPI.D4.1.CALD')
              ) %>%
  left_join(country_list)

write_excel_csv(D4.1.SC.DPO.CALD_wide,
                path = paste(csv_output, "D4.1.SC.DPO.CALD_indicator.csv", sep="/" ))
```



###  4.2 NSO has a listing of surveys and microdata sets (or NADA)

NSO has a listing of surveys and microdata sets that can provide the necessary
data and reference for follow-up.  Upon well-defined request and procedure per
the national law and practice, users and practitioners can obtain the data
collected from the households and businesses when needed.

1 Point. Yes
0 Points. No

```{r nada}

#read in csv file.
D4.2.SC.DPO.NADA <- read_csv(file = paste(csv_dir, "D4.2.SC.DPO.NADA.csv", sep="/" )) %>%
  mutate(SPI.D4.2.NADA=case_when(
    NADA==1 ~ 1, 
    NADA==0 ~ 0, 
    TRUE ~ 0 
  ))  %>%
  select(iso3c, country, date, NADA, SPI.D4.2.NADA  ) %>%
  arrange(date, country)

D4.2.SC.DPO.NADA_wide <- D4.2.SC.DPO.NADA %>%
  pivot_wider(names_from=date,
              names_sep = "_",
              values_from=c('NADA', 'SPI.D4.2.NADA')
              ) %>%
  left_join(country_list)

write_excel_csv(D4.2.SC.DPO.NADA_wide,
                path = paste(csv_output, "D4.2.SC.DPO.NADA_indicator.csv", sep="/" ))
```

### 4.3 NSO has a data portal

This indicator refers to a web-based or stand-alone platform or tools that is
maintained by the NSO and provides access to the indicators and related
metadata in systematic way.

1 Point. Yes
0 Points. No

```{r port}

#read in csv file.
D4.3.SC.DPO.PORT <- read_csv(file = paste(csv_dir, "D4.3.SC.DPO.PORT.csv", sep="/" )) %>%
  mutate(SPI.D4.3.PORT=case_when(
    PORT==1 ~ 1, 
    PORT==0 ~ 0, 
    TRUE ~ 0 
  ))  %>%
  select(iso3c, country, date, PORT, SPI.D4.3.PORT  ) %>%
  arrange(date, country)

D4.3.SC.DPO.PORT_wide <- D4.3.SC.DPO.PORT %>%
  pivot_wider(names_from=date,
              names_sep = "_",
              values_from=c('PORT', 'SPI.D4.3.PORT')
              ) %>%
  left_join(country_list)

write_excel_csv(D4.3.SC.DPO.PORT_wide,
                path = paste(csv_output, "D4.3.SC.DPO.PORT_indicator.csv", sep="/" ))
```

### 4.4 Timeseries indicators are available for download in reusable format for free

This indicator will promote usage among the users and bring accountability to
the NSO.  There are several different formats in which a user can access the
data, such as Excel, CSV, and API etc.

1 Point. Yes
0 Points. No

```{r time}

#read in csv file.
D4.4.SC.DPO.TIME <- read_csv(file = paste(csv_dir, "D4.4.SC.DPO.TIME.csv", sep="/" )) %>%
  mutate(SPI.D4.4.TIME=case_when(
    TIME==1 ~ 1, 
    TIME==0 ~ 0, 
    TRUE ~ 0 
  ))  %>%
  select(iso3c, country, date, TIME, SPI.D4.4.TIME  ) %>%
  arrange(date, country)

D4.4.SC.DPO.TIME_wide <- D4.4.SC.DPO.TIME %>%
  pivot_wider(names_from=date,
              names_sep = "_",
              values_from=c('TIME', 'SPI.D4.4.TIME')
              ) %>%
  left_join(country_list)

write_excel_csv(D4.4.SC.DPO.TIME_wide,
                path = paste(csv_output, "D4.4.SC.DPO.TIME_indicator.csv", sep="/" ))
```


### 4.5 Metadata is available providing definition, methodology, standards or classifications for existing data series

Statistical systems must be open and transparent about their methods and
procedures and provide access to adequate metadata - detailed descriptions of
the methods and procedures used to produce the data.

1 Point. Yes
0 Points. No

```{r meta}

#read in csv file.
D4.5.SC.DPO.META <- read_csv(file = paste(csv_dir, "D4.5.SC.DPO.META.csv", sep="/" )) %>%
  mutate(SPI.D4.5.META=case_when(
    META==1 ~ 1, 
    META==0 ~ 0, 
    TRUE ~ 0 
  ))  %>%
  select(iso3c, country, date, META, SPI.D4.5.META  ) %>%
  arrange(date, country)

D4.5.SC.DPO.META_wide <- D4.5.SC.DPO.META %>%
  pivot_wider(names_from=date,
              names_sep = "_",
              values_from=c('META', 'SPI.D4.5.META')
              ) %>%
  left_join(country_list)

write_excel_csv(D4.5.SC.DPO.META_wide,
                path = paste(csv_output, "D4.5.SC.DPO.META_indicator.csv", sep="/" ))
```

### 4.6 NSO has conducted a user satisfaction survey

Through this indicator a NSO can improve its engagement and dissemination
practice by seeking regular feedback from users and then making necessary
adjustments.  It is also a good practice to continually monitor the changing
data landscape and user needs as well as evolving technologies.


1 Point. Yes
0 Points. No

```{r user}

#read in csv file.
D4.6.SC.DPO.USER <- read_csv(file = paste(csv_dir, "D4.6.SC.DPO.USER.csv", sep="/" )) %>%
  mutate(SPI.D4.6.USER=case_when(
    USER==1 ~ 1, 
    USER==0 ~ 0, 
    TRUE ~ 0 
  ))  %>%
  select(iso3c, country, date, USER, SPI.D4.6.USER  ) %>%
  arrange(date, country)

D4.6.SC.DPO.USER_wide <- D4.6.SC.DPO.USER %>%
  pivot_wider(names_from=date,
              names_sep = "_",
              values_from=c('USER', 'SPI.D4.6.USER')
              ) %>%
  left_join(country_list)

write_excel_csv(D4.6.SC.DPO.USER_wide,
                path = paste(csv_output, "D4.6.SC.DPO.USER_indicator.csv", sep="/" ))
```


### 4.7 Geospatial data available on relevant agency website

Geospatial data available on relevant agency website refers to geo-referenced,
location-based disaggregated data from satellite.  Geospatial data are often
used in combination with administrative level national and sub-national units.


1 Point. Yes
0 Points. No

```{r geos}

#read in csv file.
D4.7.SC.DPO.GEOS <- read_csv(file = paste(csv_dir, "D4.7.SC.DPO.GEOS.csv", sep="/" )) %>%
  mutate(SPI.D4.7.GEOS=case_when(
    GEOS==1 ~ 1, 
    GEOS==0 ~ 0, 
    TRUE ~ 0 
  ))  %>%
  select(iso3c, country, date, GEOS, SPI.D4.7.GEOS  ) %>%
  arrange(date, country)

D4.7.SC.DPO.GEOS_wide <- D4.7.SC.DPO.GEOS %>%
  pivot_wider(names_from=date,
              names_sep = "_",
              values_from=c('GEOS', 'SPI.D4.7.GEOS')
              ) %>%
  left_join(country_list)

write_excel_csv(D4.7.SC.DPO.GEOS_wide,
                path = paste(csv_output, "D4.7.SC.DPO.GEOS_indicator.csv", sep="/" ))
```



### DPO Score

DPO Country  Score=Weighted Score÷Maximum Category Score×100

```{r dpo}
# join datasets together based on country and date
D4.DPO <- D4.1.SC.DPO.CALD %>%
  left_join(D4.2.SC.DPO.NADA) %>%
  left_join(D4.3.SC.DPO.PORT) %>%
  left_join(D4.4.SC.DPO.TIME) %>%
  left_join(D4.5.SC.DPO.META) %>%
  left_join(D4.6.SC.DPO.USER) %>%
  left_join(D4.7.SC.DPO.GEOS) %>%
  select(iso3c, country, date, starts_with("SPI.D4")) 

#Now calculate MSC score which is the average across the 12 indicators
D4.DPO <- D4.DPO %>%
  mutate(SPI.D4.DPO=100*rowMeans(.[grep(x=colnames(D4.DPO), 
                                              pattern="SPI.D4")], na.rm=TRUE)) %>%
  arrange(date, iso3c) %>%
  select(iso3c, country, date,SPI.D4.DPO, starts_with("SPI.D4"))


D4.DPO_wide <- D4.DPO %>%
  select(iso3c, country, date,SPI.D4.DPO) %>%
  pivot_wider(names_from=date,
              names_prefix = "SPI.D4.DPO_",
              values_from=c('SPI.D4.DPO')
              ) %>%
  left_join(country_list)

write_excel_csv(D4.DPO_wide,
                path = paste(csv_output, "D4.DPO_indicator.csv", sep="/" ))

```

Produce summary statistics of our indicators.

```{r}

msc_sumstats <- D4.DPO %>%
  group_by(date) %>%
  select(SPI.D4.DPO, starts_with("SPI.D4")) %>%
  skim() %>%
  select(c("skim_variable",
           "date", "complete_rate", "numeric.mean",  "numeric.sd",    "numeric.p0",    "numeric.p25",   "numeric.p50",
           "numeric.p75",   "numeric.p100",  "numeric.hist"))

kable(msc_sumstats, caption="Summary Statistics of Dissemination Practices and Openness (DPO) Indicator Variables" , col.names = c("Indicator", "Date", "Completion Rate", "Mean",  "Std Dev","Min", "25th Percentile", "Median", "75th Percentile", "Max",  "Histogram"), digits = 2 ) %>%
  kable_styling(bootstrap_options = c("striped", "hover"))

```



## Overall Score

Total SPI Score =(MSC+CS+DPO+AKI)/4

```{r spi}

#bring all the databases together
SPI<-D1.MSC %>%
  left_join(D2.CS) %>%
  left_join(D3.AKI) %>%
  left_join(D4.DPO) 

#Update ISO codes
SPI <- SPI %>%
  select(-iso3c) %>%
  left_join(country_list)



#add in country population as last step
# make request to World Bank API
populationRequest <- GET(url = "https://api.worldbank.org/v2/country/all/indicator/SP.POP.TOTL?source=40&per_page=5000&date=2015:2020&format=json")
populationResponse <- content(populationRequest, as = "text", encoding = "UTF-8")

# Parse the JSON content and convert it to a data frame.
pop <- jsonlite::fromJSON(populationResponse, flatten = TRUE) %>%
  data.frame() %>%
  mutate(population=value,
         iso3c=country.id,
         country=country.value,
         date=as.numeric(date)) %>%
  select(iso3c, country, date, population)


#merge population onto SPI
SPI <- SPI %>%
  left_join(pop)

#Construct overall score
SPI <- SPI %>%
  mutate(SPI.OVRL.SCR=(SPI.D1.MSC+SPI.D2.CS+SPI.D3.AKI+SPI.D4.DPO)/4)

#create version of data in wide format
SPI_wide <- SPI %>%
  select(iso3c, country, date, SPI.OVRL.SCR, SPI.D1.MSC, SPI.D2.CS, SPI.D3.AKI, SPI.D4.DPO) %>%
  pivot_wider(names_from=date,
              names_sep = "_",
              values_from=c('SPI.OVRL.SCR', 'SPI.D1.MSC', 'SPI.D2.CS', 'SPI.D3.AKI', 'SPI.D4.DPO')
              )

#label columns
var.labels=c(iso3c='3 digit country code', 
             country='Country Name', 
             date='Date', 
             SPI.D1.MSC='Dimension 1: Methodology, Standards & Classifications (MSC)', 
             SPI.D1.1.SNAU='Dimension 1: System of National Accounts in use', 
             SPI.D1.2.NABY='Dimension 1: National Accounts base year ', 
             SPI.D1.3.CNIN='Dimension 1: Classification of national industry', 
             SPI.D1.4.CPIBY='Dimension 1: CPI base year', 
             SPI.D1.5.HOUS='Dimension 1: Classification of household consumption', 
             SPI.D1.6.EMPL='Dimension 1: Classification of status of employment', 
             SPI.D1.7.CGOV='Dimension 1: Central government accounting status', 
             SPI.D1.8.FINA='Dimension 1: Compilation of government finance statistics', 
             SPI.D1.9.MONY='Dimension 1: Compilation of monetary and financial statistics', 
             SPI.D1.10.IDDS='Dimension 1: SDDS/e-GDDS subscription', 
             SPI.D1.11.CRVS='Dimension 1: CRVS', 
             SPI.D1.12.GSBP='Dimension 1: Business process', 
             SPI.D2.CS='Dimension 2: Censuses and Surveys (CS)', 
             SPI.D2.1.POPU='Dimension 2: Population & Housing census', 
             SPI.D2.2.AGRI='Dimension 2: Agriculture census', 
             SPI.D2.3.BIZZ='Dimension 2: Business/establishment census', 
             SPI.D2.4.HOUS='Dimension 2: Household Survey on income/consumption/expenditure/budget/Integrated Survey ', 
             SPI.D2.5.AGRI='Dimension 2: Agriculture survey', 
             SPI.D2.6.LABR='Dimension 2: Labor Force Survey', 
             SPI.D2.7.HLTH='Dimension 2: Health/Demographic survey', 
             SPI.D2.8.BIZZ='Dimension 2: Business/establishment survey', 
             SPI.D3.AKI='Dimension 3: Availability of Key Indicators (AKI)', 
             SPI.D3.POV='Dimension 3: Poverty headcount ratio at $1.90 a day (2011 PPP) (% of population)',
             SPI.D3.FIES='Dimension 3: Food Insecurity Experience Scale',
             SPI.D3.CHLD.MORT='Dimension 3: Mortality rate, under-5 (per 1,000 live births)', 
             SPI.D3.SE.LPV.PRIM.BMP='Dimension 3: Pupils below minimum reading proficiency at end of primary (%). Low GAML threshold',
             SPI.D3.MMRT='Dimension 3: Maternal Mortality',
             SPI.D3.SH.H2O.SMDW.ZS='Dimension 3: People using safely managed drinking water services (% of population)',
             SPI.D3.ELEC='Dimension 3: Access to electricity (% of population)',
             SPI.D3.SL.UEM.TOTL.NE.ZS='Dimension 3: Unemployment, total (% of total labor force) (national estimate)' ,
             SPI.D3.NV.IND.MANF.ZS='Dimension 3: Manufacturing, value added (% of GDP)',
             SPI.D3.SI.SPR.PC40.ZG='Dimension 3: Annualized average growth rate in per capita real survey mean consumption or income, bottom 40% of population (%)',
             SPI.D3.ER.H2O.FWST.ZS='Dimension 3: Level of water stress: freshwater withdrawal as a proportion of available freshwater resources',
             SPI.D3.EG.FEC.RNEW.ZS='Dimension 3: Renewable energy consumption (% of total final energy consumption)',
             SPI.D3.NE.CON.PRVT.CN='Dimension 3: Households and NPISHs Final consumption expenditure (current LCU)',
             SPI.D3.QUART.GDP='Dimension 3: Quarterly GDP',
             SPI.D3.DT.TDS.DPPF.XP.ZS='Dimension 3: Debt service (PPG and IMF only, % of exports of goods, services and primary income)', 
             SPI.D4.DPO='Dimension 4: Dissemination Practices & Openness (DPO)', 
             SPI.D4.1.CALD='Dimension 4: NSO has an Advance Release Calendar and it is published ', 
             SPI.D4.2.NADA='Dimension 4: NSO has a listing of surveys and microdata sets (or NADA)', 
             SPI.D4.3.PORT='Dimension 4: NSO has a data portal', 
             SPI.D4.4.TIME='Dimension 4: Timeseries indicators are available for download in reusable format for free', 
             SPI.D4.5.META='Dimension 4: Metadata is available providing definition, methodology, standards or classifications for existing data series', 
             SPI.D4.6.USER='Dimension 4: NSO has conducted a user satisfaction survey', 
             SPI.D4.7.GEOS='Dimension 4: Geospatial data available on relevant agency website', 
             SPI.OVRL.SCR='SPI Overall Score')

#Order SPI data
SPI <- SPI %>%
  select(country, iso3c, date, names(var.labels), everything())


#label data
SPI_unlabelled <- SPI #create unlabelled version to save
label(SPI) = as.list(var.labels[match(names(SPI), names(var.labels))])


#turn variable labels into data frame 
var.labels.df<-data.frame(var.labels) %>% 
  rownames_to_column() %>%
  rename(variable=rowname)

#export data
write_excel_csv(SPI,
                path = paste(csv_output, "SPI_scores_long.csv", sep="/" ))

SPI_labelled <- SPI %>%
    filter((lending %in% c('IDA', 'IBRD', 'Blend')) | (iso3c=='PSE')) #keep just the borrowing countries & West Bank & Gaza

colnames(SPI_labelled) = as.list(var.labels[match(names(SPI), names(var.labels))])
write_excel_csv(SPI_labelled,
                path = paste(csv_output, "SPI_scores_long_labelled.csv", sep="/" ))

# #save to stata with compatible names
 SPI_stata <- SPI  %>%
   filter((lending %in% c('IDA', 'IBRD', 'Blend')) | (iso3c=='PSE')) %>% #keep just the borrowing countries & West Bank & Gaza
   select(country, iso3c,date, contains("SPI.D"), contains("SPI.O"), region, regionID, income, incomeID, lending, lendingID ) %>%
   janitor::clean_names() 

write_dta(SPI_stata, path = paste(csv_output, "SPI_scores.dta", sep="/" ), version=15)


write_excel_csv(SPI_wide,
                path = paste(csv_output, "SPI_scores_wide.csv", sep="/" ))

# SPI_wide  %>%
#   janitor::clean_names() %>%
#   write_dta(path = paste(csv_output, "SPI_scores_wide.dta", sep="/" ))
```

# Metadata

```{r metadata}

#turn variable labels into data frame for metadata
metadata <- read_csv(file=paste(csv_dir, "SPI_metadata.csv", sep="/"))

kable(metadata, caption="Metadata for Statistical Performance Index Indicators" ) %>%
  kable_styling(bootstrap_options = c("striped", "hover")) %>%
  column_spec(1:5, width_max  = "20em")

```


```{r save_R_frame}

save(SPI, SPI_wide, metadata,
     file = paste(csv_output, "SPI.Rdata", sep="/" ))

save(SPI_unlabelled, SPI_wide, metadata,
     file = paste(csv_output, "SPI_unlabelled.Rdata", sep="/" ))
```
