==> R/04_WDR/042_programs/03_SPI_tables_for_WDR.Rmd <==
---
title: "Statistical Performance Indicators - Special Requests for WDR"
author: "Brian Stacy"
date: "`r Sys.Date()`"
output:
  html_document: 
    toc: yes
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo=FALSE, warning = FALSE, message = FALSE)

# Load Packages
library(tidyverse)
library(skimr)
library(kableExtra)
library(readxl)
library(Hmisc)
library(haven)
library(gt)
library(xlsx)

# Directory for SPI csv files that were previously created
csv_inputs <- "C:/Users/wb469649/OneDrive - WBG/Documents/Github/SPI_AKI/R/01_data/011_rawdata"
csv_dir <- "C:/Users/wb469649/OneDrive - WBG/Documents/Github/SPI_AKI/R/01_data/013_outputs"
out_dir <- "C:/Users/wb469649/OneDrive - WBG/Documents/Github/SPI_AKI/R/04_WDR/043_outputs"

country_metadata <- wbstats::wbcountries() %>%
  filter(region!='Aggregates')

```

# Special Requests for WDR

This file will clean the raw SPI data and form a set of indicators for use in the WDR.  

Hi Brian, 

Thanks so much for sending this my way – super super interesting materials. I am looking forward to seeing how the discussion with John Pullinger will evolve, and I’m sure I will be interested in citing some of the additional indicators. 
In the meantime, I have updated the excel document as discussed. It is quite a long list of requests, please don’t hesitate to let me know if this is too much of a burden on your time. I have put indicator breakdowns as income level (LIC, LMC, UMC, HIC) and WGB regions, though we probably won’t be citing regions very often so if it’s a lot of extra work, feel free to disregard that. On the issue of weighting, do you recommend population weighting or not? 

On the issue of child/infant mortality and maternal mortality, could you share with me the information as % of countries with vital reg source, survey source, no source as well as by each of the frequency intervals, broken down by income level? Again, please point out if too much work! 

On povcalnet, I am wondering what kind of information you have on comparability of the data over time? 

Again, thanks so much, Brian, for all your help with this, and please do let me know if this is asking too much!  
All the best,
Philip


## Table

| Section | Indicator | Source | SPI Inidicator | Proposed indicator |
|-|-|-|-|-|
| Coverage | coverage/availability score | ODIN | Could use SPI main score | Scores for dimensions 2 and 3 (separately), by income level and region |
| Coverage | Poverty since 2015 | Povcalnet | Also in SPI, but Povcalnet is   our source | See email |
| Coverage | Monthly or quarterly industrial production indices by income group | SCI | Quarterly GDP is included in SPI | % of countries with 3 or more values of quarterly GDP by incomelevel and   region; % of countries with no values in past 5 years by income level and   region |
| Coverage | Half of agricultural data needs met in SSA | GSARS | We have data on all agricultural   census or agriculture surveys in past 10 years | will drop this from text |
| Coverage | Gender statistics availability | UN Women | Gender indicator is Maternal   Mortality, but have been working to build database on whether breakdowns by   gender are available for several of first 10 SDGs | see email  |
| Coverage | PH Census in last 10 years | SCI | SPI as well, which we can   customize | % of countries with PH census in last ten years by income level and   region |
| Coverage | Ag Census in last 10 years | SCI | SPI as well, which we can   customize | same as PH census indicator |
| Coverage | Complete vital stats coverage | SCI  | SPI as well, which we can   customize | % of complete CRVS by income and region |
| Coverage | Number of people uncounted | UN Women | We do not have this | will use original source |
| Coverage | Geospatial: addresses | Global Open Data Index | We do not have this | will use original source |
| Coverage | Geospatial: maps | Global Open Data Index | We do not have this | will use original source |
| Coverage | Geospatial: national maps | Global Open Data Index | We do not have this | will use original source |
| Coverage | GDP per capita | SCI | We have actually dropped GDP per   capita, because we decided to include Quarterly GDP | use quarterly GDP, see above |
| Coverage | Child mortality | SCI | SPI has good child mortality   data | let's use, see email |
| Coverage | Immunization rates | SCI | Do not have | will drop this from text |
| Coverage | Disaggregation at admin1 | ODIN | Do not have | will use original source |
| Coverage | Disaggregation at admin2 | ODIN | Do not have | will use original source |
| Openness & Interoperability | Data Openness Score | ODIN | We have a subscore for data   dissemination practices and openness | Dimension 4 score by income level and region |
| Openness & Interoperability | Open terms of use agreement  | ODIN | Do not have | will be dropped or used from original source |
| Openness & Interoperability | Available in machine-readable format  | ODIN | Do not have | will be dropped or used from original source |
| Capacity | SCI score | SCI | We have this | (preliminary) SPI score by income level |
| Governance | NSO independence | Ibrahim Foundation | Do not have | will use original source |
| Additional   Indicators |  |  |  |  |
|  |  |  |  | 4.1, 4.2, 4.3, 4.5, 4.6, 4.7 by income level and region |
|  |  |  |  | Manufacturing, value added (% of GDP), % of countries without any in past   5 years by income and region |
|  |  |  |  | GNI current LCU, % of countries without any in past 5 years by income and   region |






```{r load1, include=FALSE}

#import data
load(paste(out_dir, "SPI_unlabelled.RData", sep="/" ))

SPI <- SPI_unlabelled %>%
  filter((lending %in% c('IDA', 'IBRD', 'Blend')) | (iso3c=='PSE')) #keep just the borrowing countries & West Bank & Gaza

```

# % of countries with 3 or more values of quarterly GDP by incomelevel and region; % of countries with no values in past 5 years by income level and region


Quarterly GDP numbers were pulled from the IMF website.

After GDP by expenditure, quarterly GDP is probably the most important development to be made in a system of National Accounts, before the development of full sectoral accounts. IMF IFS has quarterly GDP from various sources, including governments and international agencies at https://data.imf.org/?sk=4C514D48-B6BA-49ED-8AB9-52B0C1A0179B




```{r AKI_3.17}

#read in file from rawdata folder
D3.14.QUART.GDP <- read_csv(file = paste(csv_inputs, "D3.14.QUART.GDP.csv", sep="/" ))

#clean data and produce indicator for each year
#Now loop from 2016 and 2019, keeping just data inside last 5 years.
reference_year <-2019

QGDP <- D3.14.QUART.GDP %>%
  mutate(date=as.numeric(str_sub(TIME_PERIOD,1,4))) %>%
  mutate(iso2c=REF_AREA) %>%
  mutate(frequency=((reference_year-as.numeric(date))<=5) & (reference_year>=as.numeric(date))) %>% 
  mutate(SPI.D3.QUART.GDP=if_else(frequency==TRUE,1,0)) %>% #create 0,1 variable for whether data point exists for country
  group_by(iso2c, date) %>%
  summarise(SPI.D3.QUART.GDP=max(SPI.D3.QUART.GDP, na.rm=T)) %>% 
  group_by(iso2c) %>%
  summarise(SPI.D3.QUART.GDP=sum(SPI.D3.QUART.GDP, na.rm=T)) %>% 
  mutate(SPI.D3.QUART.GDP=case_when(
    SPI.D3.QUART.GDP>=3 ~ '3 or More values of Quarterly GDP in Past 5 Years',
    #SPI.D3.QUART.GDP==2 ~ 0.6,
    #SPI.D3.QUART.GDP==1 ~ 0.3,
    SPI.D3.QUART.GDP==0 ~ 'No values of Quarterly GDP in Past 5 Years', 
    TRUE ~ '1-2 Values in Past 5 Years'
  )) %>%
  mutate(date=reference_year) %>%
  select( iso2c, date, starts_with('SPI.D3.QUART.GDP')) %>%
  right_join(country_metadata) %>%
  mutate_at(vars(starts_with('SPI.D3.')), ~if_else(is.na(.),'No values of Quarterly GDP in Past 5 Years',.)) 


#table by region

tab <- QGDP %>%
  group_by(region) %>%
  mutate(total=n()) %>%
  group_by(region, SPI.D3.QUART.GDP) %>%
  summarise(N=n(),
            total=mean(total, na.rm=T)) %>%
  mutate(Percent = 100*round(N/total,3))   %>%
  select(region,SPI.D3.QUART.GDP, Percent) %>%
  pivot_wider(names_from=SPI.D3.QUART.GDP,
              values_from='Percent',
              values_fill=0)

tab_ovrl <- QGDP %>%
  ungroup() %>%
  mutate(total=n()) %>%
  group_by( SPI.D3.QUART.GDP) %>%
  summarise(N=n(),
            total=mean(total, na.rm=T)) %>%
  mutate(region="Global") %>%
  mutate(Percent = 100*round(N/total,3))   %>%
  select(region,SPI.D3.QUART.GDP, Percent) %>%
  pivot_wider(names_from=SPI.D3.QUART.GDP,
              values_from='Percent',
              values_fill=0)


  kable(bind_rows(tab, tab_ovrl), caption='% of Countries by Number of Quarterly GDP Values by Region') %>%
  kable_styling()
  

#table by income

tab_inc <- QGDP %>%
  group_by(income) %>%
  mutate(total=n()) %>%
  group_by(income, SPI.D3.QUART.GDP) %>%
  summarise(N=n(),
            total=mean(total, na.rm=T)) %>%
  mutate(Percent = 100*round(N/total,3))   %>%
  select(income,SPI.D3.QUART.GDP, Percent) %>%
  pivot_wider(names_from=SPI.D3.QUART.GDP,
              values_from='Percent',
              values_fill=0)

tab_ovrl_inc <- QGDP %>%
  ungroup() %>%
  mutate(total=n()) %>%
  group_by( SPI.D3.QUART.GDP) %>%
  summarise(N=n(),
            total=mean(total, na.rm=T)) %>%
  mutate(income="Global") %>%
  mutate(Percent = 100*round(N/total,3))   %>%
  select(income,SPI.D3.QUART.GDP, Percent) %>%
  pivot_wider(names_from=SPI.D3.QUART.GDP,
              values_from='Percent',
              values_fill=0)


  kable(bind_rows(tab_inc, tab_ovrl_inc), caption='% of Countries by Number of Quarterly GDP Values by Income Group') %>%
  kable_styling()
  


```
# % of countries with vital reg source, survey source, no source as well as by each of the frequency intervals, broken down by income level

##  Mortality rate, under-5 (per 1,000 live births)  

The data in WDI is modeled, based on HHS and Vital Registration.  We use the following as a source for raw data produced by national statistical offices that is used in this modeling.  

https://childmortality.org/data 

Data was pulled on April 13, 2020.




```{r aki_3.3, echo=FALSE}

#Read in data from UN Inter-agency Group for Child Mortality Estimation

D3.3.AKI.MORT <- read_csv(file=paste(csv_inputs, 'D3.3_child_mort_RAW_v01.csv', sep="/")) %>%
  filter(INDICATOR=='Under-five mortality rate' & SEX=='Total') %>% #keep just observations for under 5 child mortality and for both sexes 
  filter(OBS_STATUS=='Included in IGME') %>% # Also keep only surveys that met IGME criteria for inclusion as a nationally representative statistic
  mutate(date=as.numeric(str_extract(TIME_PERIOD, "^.{4}")),
         country=REF_AREA,
         D3.CHLD.MORT=OBS_VALUE)

#Now loop from 2016 and 2019, keeping just data inside last 5 years.
reference_year<-2019
  MORT <-D3.3.AKI.MORT %>%
          mutate(frequency=((reference_year-as.numeric(date))<=5) & (reference_year>=as.numeric(date))) %>% 
          mutate(SPI.FREQ.D3.CHLD.MORT=if_else(frequency==TRUE,1,0)) %>% #create 0,1 variable for whether data point exists for country
          mutate(SPI.QUAL.D3.CHLD.MORT=case_when(
                   reference_year-as.numeric(date)>=5 ~ 0, #outside 5 year window
                   SERIES_METHOD=='Vital Registration' ~ 1,
                  SERIES_METHOD!='Vital Registration' ~ 0.5,
                   TRUE ~ 0
                 )) %>%
          group_by(country) %>%
          summarise(SPI.FREQ.D3.CHLD.MORT=sum(SPI.FREQ.D3.CHLD.MORT, na.rm=T),
                    SPI.QUAL.D3.CHLD.MORT=max(SPI.QUAL.D3.CHLD.MORT, na.rm=T)) %>% 
          mutate(SPI.QUAL.D3.CHLD.MORT=case_when(
                   SPI.QUAL.D3.CHLD.MORT==1 ~ 'Based on Vital Statistics Record',
                   SPI.QUAL.D3.CHLD.MORT==0.5 ~ 'Survey Source',
                   TRUE ~ 'No Source'
                 ))  %>%
          mutate(date=reference_year) %>%
          mutate(country=case_when(
            country=="Bahamas" ~ "Bahamas, The",                  
            country=="Bolivia (Plurinational State of)"   ~  "Bolivia"   ,                  
            country=="Côte d'Ivoire"  ~  "Cote d'Ivoire"                ,
            country=="Democratic Republic of the Congo"  ~ "Congo, Dem. Rep."  ,           
            country=="Congo"  ~ "Congo, Rep."        ,           
            country=="Curacao"  ~  "Curacao"       ,                
            country=="Czechia"   ~  "Czech Republic"     ,        
            country=="Egypt"  ~ "Egypt, Arab Rep."        ,      
            country=="Micronesia (Federated States of)"  ~ "Micronesia, Fed. Sts."  ,       
            country=="United Kingdom"   ~  "United Kingdom"    ,         
            country=="Gambia"    ~  "Gambia, The"     ,           
            country=="Iran (Islamic Republic of)"  ~ "Iran, Islamic Rep." ,           
            country=="Kyrgyzstan"   ~  "Kyrgyz Republic"   ,          
            country=="Republic of Korea"   ~ "Korea, Rep."   ,               
            country=="Lao People's Democratic Republic"    ~  "Lao PDR" ,                    
            country=="Saint Kitts and Nevis	"   ~ "St. Kitts and Nevis",
            country=="Saint Lucia"   ~ "St. Lucia",
            country=="Republic of Moldova"   ~  "Moldova"   ,                  
            country=="Democratic People's Republic of Korea"  ~  "Korea, Dem. People’s Rep." ,  
            country=="Slovakia"   ~  "Slovak Republic"             ,
            country=="United Republic of Tanzania"   ~  "Tanzania" ,                    
            country=="United States of America"  ~ "United States"  ,                
            country=="Saint Vincent and the Grenadines" ~ "St. Vincent and the Grenadines" ,
            country=="Venezuela (Bolivarian Republic of)"  ~  "Venezuela, RB"   ,              
            country=="British Virgin Islands"  ~  "British Virgin Islands" ,       
            country=="Viet Nam"  ~ "Vietnam"        ,              
            country=="Yemen"  ~ "Yemen, Rep.",
            country=="United Kingdom of Great Britain and Northern Ireland" ~ "United Kingdom",
            TRUE ~ country
          )) %>%

          select( country, date, starts_with('SPI.QUAL.D3.CHLD.MORT')) %>%
          right_join(country_metadata) %>%
          mutate_at(vars(starts_with('SPI.QUAL.D3.CHLD.MORT')), ~if_else(is.na(.),'No Source',.)) 




#table by region

tab <- MORT %>%
  group_by(region) %>%
  mutate(total=n()) %>%
  group_by(region, SPI.QUAL.D3.CHLD.MORT) %>%
  summarise(N=n(),
            total=mean(total, na.rm=T)) %>%
  mutate(Percent = 100*round(N/total,3))   %>%
  select(region,SPI.QUAL.D3.CHLD.MORT, Percent) %>%
  pivot_wider(names_from=SPI.QUAL.D3.CHLD.MORT,
              values_from='Percent',
              values_fill=0)

tab_ovrl <- MORT %>%
  ungroup() %>%
  mutate(total=n()) %>%
  group_by( SPI.QUAL.D3.CHLD.MORT) %>%
  summarise(N=n(),
            total=mean(total, na.rm=T)) %>%
  mutate(region="Global") %>%
  mutate(Percent = 100*round(N/total,3))   %>%
  select(region,SPI.QUAL.D3.CHLD.MORT, Percent) %>%
  pivot_wider(names_from=SPI.QUAL.D3.CHLD.MORT,
              values_from='Percent',
              values_fill=0)


  kable(bind_rows(tab, tab_ovrl), caption='% of Countries by Child Mortality Source by Region from 2014-2019') %>%
  kable_styling()
  

#table by income

tab_inc <- MORT %>%
  group_by(income) %>%
  mutate(total=n()) %>%
  group_by(income, SPI.QUAL.D3.CHLD.MORT) %>%
  summarise(N=n(),
            total=mean(total, na.rm=T)) %>%
  mutate(Percent = 100*round(N/total,3))   %>%
  select(income,SPI.QUAL.D3.CHLD.MORT, Percent) %>%
  pivot_wider(names_from=SPI.QUAL.D3.CHLD.MORT,
              values_from='Percent',
              values_fill=0)

tab_ovrl_inc <- MORT %>%
  ungroup() %>%
  mutate(total=n()) %>%
  group_by( SPI.QUAL.D3.CHLD.MORT) %>%
  summarise(N=n(),
            total=mean(total, na.rm=T)) %>%
  mutate(income="Global") %>%
  mutate(Percent = 100*round(N/total,3))   %>%
  select(income,SPI.QUAL.D3.CHLD.MORT, Percent) %>%
  pivot_wider(names_from=SPI.QUAL.D3.CHLD.MORT,
              values_from='Percent',
              values_fill=0)


  kable(bind_rows(tab_inc, tab_ovrl_inc), caption='% of Countries by Child Mortality Source by Income Group from 2014-2019') %>%
  kable_styling()


```


## Maternal Mortality Ratio

The data is sourced from the the Inter-Agency Group (200 countries) to improve time and country coverages. Tne national estimates with the source information are available at https://www.who.int/reproductivehealth/publications/maternal-mortality-2000-2017/en/.  The data was accessed on April 3, 2020.  This data was used for modelling maternal mortality, which are then fed into the modelled estimates on Maternal Mortality (SH.STA.MMRT) in the WDI.




```{r AKI_3.5, echo=FALSE}


#Read in data from UN Inter-agency Group 

D3.5.AKI.MMR <- read_csv(file=paste(csv_inputs, 'D3.5.BMat2019_datainputs.csv', sep="/")) %>%
  filter(modelinclude=='TRUE') %>% #keep just observations that the UN inter-agency group deemed credible for use in modelling 
  mutate(date=end,
         iso3c=iso,
         D3.MMRT=pm_obs)

#Now loop from 2016 and 2019, keeping just data inside last 5 years.
reference_year<-2019
  MMR <-D3.5.AKI.MMR %>%
          mutate(frequency=((reference_year-as.numeric(date))<=5) & (reference_year>=as.numeric(date))) %>% 
          mutate(SPI.FREQ.D3.MMRT=if_else(frequency==TRUE,1,0)) %>% #create 0,1 variable for whether data point exists for country
          mutate(SPI.QUAL.D3.MMRT=case_when(
                   reference_year-as.numeric(date)>=5 ~ 0, #outside 5 year window
                   type=='vr' ~ 1,
                   type!='vr' ~ 0.5,
                   TRUE ~ 0
                 )) %>%
          group_by(iso3c) %>%
          summarise(SPI.FREQ.D3.MMRT=sum(SPI.FREQ.D3.MMRT, na.rm=T),
                    SPI.QUAL.D3.MMRT=max(SPI.QUAL.D3.MMRT, na.rm=T)) %>%     
          mutate(SPI.QUAL.D3.MMRT=case_when(
                   SPI.QUAL.D3.MMRT==1 ~ 'Based on Vital Statistics Record',
                   SPI.QUAL.D3.MMRT==0.5 ~ 'Survey Source',
                   TRUE ~ 'No Source'
                 ))  %>%
          mutate(date=reference_year) %>%
          select( iso3c, date, starts_with('SPI.QUAL.D3.MMRT')) %>%
          right_join(country_metadata) %>%
          mutate_at(vars(starts_with('SPI.QUAL.D3.MMRT')), ~if_else(is.na(.),'No Source',.)) 

          



#table by region

tab <- MMR %>%
  group_by(region) %>%
  mutate(total=n()) %>%
  group_by(region, SPI.QUAL.D3.MMRT) %>%
  summarise(N=n(),
            total=mean(total, na.rm=T)) %>%
  mutate(Percent = 100*round(N/total,3))   %>%
  select(region,SPI.QUAL.D3.MMRT, Percent) %>%
  pivot_wider(names_from=SPI.QUAL.D3.MMRT,
              values_from='Percent',
              values_fill=0)

tab_ovrl <- MMR %>%
  ungroup() %>%
  mutate(total=n()) %>%
  group_by( SPI.QUAL.D3.MMRT) %>%
  summarise(N=n(),
            total=mean(total, na.rm=T)) %>%
  mutate(region="Global") %>%
  mutate(Percent = 100*round(N/total,3))   %>%
  select(region,SPI.QUAL.D3.MMRT, Percent) %>%
  pivot_wider(names_from=SPI.QUAL.D3.MMRT,
              values_from='Percent',
              values_fill=0)


  kable(bind_rows(tab, tab_ovrl), caption='% of Countries by Maternal Mortality Source by Region from 2014-2019') %>%
  kable_styling()
  

#table by income

tab_inc <- MMR %>%
  group_by(income) %>%
  mutate(total=n()) %>%
  group_by(income, SPI.QUAL.D3.MMRT) %>%
  summarise(N=n(),
            total=mean(total, na.rm=T)) %>%
  mutate(Percent = 100*round(N/total,3))   %>%
  select(income,SPI.QUAL.D3.MMRT, Percent) %>%
  pivot_wider(names_from=SPI.QUAL.D3.MMRT,
              values_from='Percent',
              values_fill=0)

tab_ovrl_inc <- MMR %>%
  ungroup() %>%
  mutate(total=n()) %>%
  group_by( SPI.QUAL.D3.MMRT) %>%
  summarise(N=n(),
            total=mean(total, na.rm=T)) %>%
  mutate(income="Global") %>%
  mutate(Percent = 100*round(N/total,3))   %>%
  select(income,SPI.QUAL.D3.MMRT, Percent) %>%
  pivot_wider(names_from=SPI.QUAL.D3.MMRT,
              values_from='Percent',
              values_fill=0)


  kable(bind_rows(tab_inc, tab_ovrl_inc), caption='% of Countries by Maternal Mortality Source by Income Group from 2014-2019') %>%
  kable_styling()
```





# Overall SPI scores by Region and Income Level

This file produces a number of summary statistics for the SPI dimensions.  Summary statistics are presented both population weighted and unweighted.  We will produce a table of summary statistics with mean, median, standard deviation, etc, and also a table of means by region and income level.


```{r load, include=FALSE}

#import data
load(paste(csv_dir, "SPI_unlabelled.RData", sep="/" ))

SPI <- SPI_unlabelled %>%
  filter((lending %in% c('IDA', 'IBRD', 'Blend')) | (iso3c=='PSE')) #keep just the borrowing countries & West Bank & Gaza

```



# Summary Statistics

Below we show summary statistics for the year 2019.  The same table is available for other years in the appendix.  The first table shows the summary statistics weighted by population.  The second shows unweighted.

## Population Weighted Summary Statistics

```{r sumstats_weighted, echo=FALSE}

#add function to produce weighted summary stats
        my_skim<-    skim_with( numeric = sfl( mean = ~ wtd.mean(.,  weights =pop, na.rm=TRUE),
                                               sd = ~ sqrt(wtd.var(.,  weights=pop, na.rm=TRUE)),
                                               p25 = ~ (wtd.quantile(., probs=c(0.25),  weights=pop, na.rm=TRUE)),
                                               p50 = ~ (wtd.quantile(., probs=c(0.5), weights=pop, na.rm=TRUE)),
                                               p75 = ~ (wtd.quantile(., probs=c(0.75), weights=pop, na.rm=TRUE)),
                                               N = ~ sum(!is.na(.))))

for (reference_year in 2016:2019) {
temp <- SPI %>%
  filter(date==reference_year) 

#add population weight
pop <- temp$population

temp <- temp %>%
  select( SPI.OVRL.SCR, contains('SPI.D1'),contains('SPI.D2'),contains('SPI.D3'),contains('SPI.D4')) %>%
  my_skim() %>%
  yank('numeric') %>%
  rename(Series=skim_variable) %>%
  left_join(metadata) %>% #add labels to variables
  select(Series, 'Indicator Name', mean, sd,p0,p100, N)

assign(paste("spi_weighted_sumstats",reference_year,sep="_"), temp)

}

#save these to an excel file for easy sharing
write.xlsx(
  spi_weighted_sumstats_2016,
  file=paste(out_dir, 'summary_statistics_SPI.xlsx', sep="/"),
  sheetName='spi_weighted_sumstats_2016',
  append=FALSE
) 

write.xlsx(
  spi_weighted_sumstats_2017,
  file=paste(out_dir, 'summary_statistics_SPI.xlsx', sep="/"),
  sheetName='spi_weighted_sumstats_2017',
  append=T
)     

write.xlsx(
  spi_weighted_sumstats_2018,
  file=paste(out_dir, 'summary_statistics_SPI.xlsx', sep="/"),
  sheetName='spi_weighted_sumstats_2018',
  append=T
)     

write.xlsx(
  spi_weighted_sumstats_2019,
  file=paste(out_dir, 'summary_statistics_SPI.xlsx', sep="/"),
  sheetName='spi_weighted_sumstats_2019',
  append=T
)     



#create grouping variable for table
group <- spi_weighted_sumstats_2019 %>%
  mutate(group=case_when(
    grepl('OVRL', Series) ~ "Overall",
    grepl('D1', Series) ~ "Dimension 1",
    grepl('D2', Series) ~ "Dimension 2",
    grepl('D3', Series) ~ "Dimension 3",
    grepl('D4', Series) ~ "Dimension 4"
  ))

kable(spi_weighted_sumstats_2019,
      digits=2,
      col.names=c('Series', 'Description', 'Mean','SD','Min','Max', 'N'),
      caption="Statistical Performance Indicator Population Weighted Summary Statistics for 2019",
      format = 'html') %>%
  kable_styling(bootstrap_options = c('striped','hover')) %>%
  pack_rows(index=table(fct_inorder(group$group)))
      
      



```

## Regional Means (Weighted by Population)

```{r regional_weighted, echo=FALSE}

for (reference_year in 2016:2019) {
df_overall <- SPI %>%
  filter(date==reference_year) 
#produce by region
        sumstats<- df_overall %>%
            group_by(region) %>%
            filter(!is.na(region)) %>%
            select(region, SPI.OVRL.SCR, contains('SPI.D1'),contains('SPI.D2'),contains('SPI.D3'),contains('SPI.D4'), population) %>%
            summarise_all(~round(100*wtd.mean(.,population, na.rm=T),2)) %>%
            select(-population) %>%
            mutate(SPI.OVRL.SCR=round(SPI.OVRL.SCR/100,2),
                   SPI.D1.MSC=round(SPI.D1.MSC/100,2),
                   SPI.D2.CS=round(SPI.D2.CS/100,2),
                   SPI.D3.AKI=round(SPI.D3.AKI/100,2),
                   SPI.D4.DPO=round(SPI.D4.DPO/100,2)) 
        
        #produce global number
        sumstats_gl<- df_overall %>%
            mutate(region='Global') %>%
            group_by(region) %>%
            select(region, SPI.OVRL.SCR, contains('SPI.D1'),contains('SPI.D2'),contains('SPI.D3'),contains('SPI.D4'), population) %>%
            summarise_all(~round(100*wtd.mean(.,population, na.rm=T),2)) %>%
            select(-population) %>%
            mutate(SPI.OVRL.SCR=round(SPI.OVRL.SCR/100,2),
                   SPI.D1.MSC=round(SPI.D1.MSC/100,2),
                   SPI.D2.CS=round(SPI.D2.CS/100,2),
                   SPI.D3.AKI=round(SPI.D3.AKI/100,2),
                   SPI.D4.DPO=round(SPI.D4.DPO/100,2))
        
        
        #transpose data
        sumstats_df_long <-sumstats_gl %>%
            bind_rows(sumstats) 
        
        sumstats_df <- as.data.frame(t(sumstats_df_long %>% select(-region)))
        colnames(sumstats_df) = sumstats_df_long$region 
        
        
        sumstats_df <- sumstats_df %>%
            rownames_to_column() %>%
            rename(series=rowname)
        
        
        #create labels df
        metadata_tab2_overall <- metadata %>% 
            janitor::clean_names() %>%
            select(series, indicator_name)
        
        
        #add variable label
        sumstats_df <- sumstats_df %>%
            left_join(metadata_tab2_overall) %>%
            rename(Series=series,
                   Label=indicator_name) %>%
            select(Series, Label, everything())

 assign(paste("spi_weighted_regional",reference_year,sep="_"), sumstats_df)
       
        }
 #save these to an excel file for easy sharing
write.xlsx(
  spi_weighted_regional_2016,
  file=paste(out_dir, 'summary_statistics_SPI.xlsx', sep="/"),
  sheetName='spi_weighted_regional_2016',
  append=T
) 

write.xlsx(
  spi_weighted_regional_2017,
  file=paste(out_dir, 'summary_statistics_SPI.xlsx', sep="/"),
  sheetName='spi_weighted_regional_2017',
  append=T
) 
write.xlsx(
  spi_weighted_regional_2018,
  file=paste(out_dir, 'summary_statistics_SPI.xlsx', sep="/"),
  sheetName='spi_weighted_regional_2018',
  append=T
) 
write.xlsx(
  spi_weighted_regional_2019,
  file=paste(out_dir, 'summary_statistics_SPI.xlsx', sep="/"),
  sheetName='spi_weighted_regional_2019',
  append=T
) 
                
        DT::datatable(spi_weighted_regional_2019, caption=htmltools::tags$caption(
            style = 'caption-side: top; text-align: left;',
            htmltools::em("Population Weighted Mean of SPI Indicators by Region in 2019.  Sub-indicators scaled to be 0-100."))    ,
            extensions = 'Buttons', options=list(
                dom = 'Bfrtip',
                buttons = c('copy', 'csv', 'excel', 'pdf', 'print'),
                pageLength = 60)) 

```

## Means by Income Group (Weighted by Population)

```{r income_weighted, echo=FALSE}

for (reference_year in 2016:2019) {
df_overall <- SPI %>%
  filter(date==reference_year) 
#produce by region
        sumstats<- df_overall %>%
            group_by(income) %>%
            filter(!is.na(income)) %>%
            select(income, SPI.OVRL.SCR, contains('SPI.D1'),contains('SPI.D2'),contains('SPI.D3'),contains('SPI.D4'), population) %>%
            summarise_all(~round(100*wtd.mean(.,population, na.rm=T),2)) %>%
            select(-population) %>%
            mutate(SPI.OVRL.SCR=round(SPI.OVRL.SCR/100,2),
                   SPI.D1.MSC=round(SPI.D1.MSC/100,2),
                   SPI.D2.CS=round(SPI.D2.CS/100,2),
                   SPI.D3.AKI=round(SPI.D3.AKI/100,2),
                   SPI.D4.DPO=round(SPI.D4.DPO/100,2)) 
        
        #produce global number
        sumstats_gl<- df_overall %>%
            mutate(income='Global') %>%
            group_by(income) %>%
            select(income, SPI.OVRL.SCR, contains('SPI.D1'),contains('SPI.D2'),contains('SPI.D3'),contains('SPI.D4'), population) %>%
            summarise_all(~round(100*wtd.mean(.,population, na.rm=T),2)) %>%
            select(-population) %>%
            mutate(SPI.OVRL.SCR=round(SPI.OVRL.SCR/100,2),
                   SPI.D1.MSC=round(SPI.D1.MSC/100,2),
                   SPI.D2.CS=round(SPI.D2.CS/100,2),
                   SPI.D3.AKI=round(SPI.D3.AKI/100,2),
                   SPI.D4.DPO=round(SPI.D4.DPO/100,2))
        
        
        #transpose data
        sumstats_df_long <-sumstats_gl %>%
            bind_rows(sumstats) 
        
        sumstats_df <- as.data.frame(t(sumstats_df_long %>% select(-income)))
        colnames(sumstats_df) = sumstats_df_long$income 
        
        
        sumstats_df <- sumstats_df %>%
            rownames_to_column() %>%
            rename(series=rowname)
        
        
        #create labels df
        metadata_tab2_overall <- metadata %>% 
            janitor::clean_names() %>%
            select(series, indicator_name)
        
        
        #add variable label
        sumstats_df <- sumstats_df %>%
            left_join(metadata_tab2_overall) %>%
            rename(Series=series,
                   Label=indicator_name) %>%
            select(Series, Label, everything())

 assign(paste("spi_weighted_income",reference_year,sep="_"), sumstats_df)
       
        }
 #save these to an excel file for easy sharing
write.xlsx(
  spi_weighted_income_2016,
  file=paste(out_dir, 'summary_statistics_SPI.xlsx', sep="/"),
  sheetName='spi_weighted_income_2016',
  append=T
) 

write.xlsx(
  spi_weighted_income_2017,
  file=paste(out_dir, 'summary_statistics_SPI.xlsx', sep="/"),
  sheetName='spi_weighted_income_2017',
  append=T
) 
write.xlsx(
  spi_weighted_income_2018,
  file=paste(out_dir, 'summary_statistics_SPI.xlsx', sep="/"),
  sheetName='spi_weighted_income_2018',
  append=T
) 
write.xlsx(
  spi_weighted_income_2019,
  file=paste(out_dir, 'summary_statistics_SPI.xlsx', sep="/"),
  sheetName='spi_weighted_income_2019',
  append=T
) 
                
        DT::datatable(spi_weighted_income_2019, caption=htmltools::tags$caption(
            style = 'caption-side: top; text-align: left;',
            htmltools::em("Population Weighted Mean of SPI Indicators by Income in 2019.  Sub-indicators scaled to be 0-100."))    ,
            extensions = 'Buttons', options=list(
                dom = 'Bfrtip',
                buttons = c('copy', 'csv', 'excel', 'pdf', 'print'),
                pageLength = 60)) 

```


## Unweighted Summary Statistics

```{r sumstats_unweighted, echo=FALSE}

        my_skim<-    skim_with( numeric = sfl( mean = ~ mean(.,   na.rm=TRUE),
                                               sd = ~ sqrt(var(.,   na.rm=TRUE)),
                                               p25 = ~ (quantile(., probs=c(0.25),   na.rm=TRUE)),
                                               p50 = ~ (quantile(., probs=c(0.5),  na.rm=TRUE)),
                                               p75 = ~ (quantile(., probs=c(0.75),  na.rm=TRUE)),
                                               N = ~ sum(!is.na(.))))

for (reference_year in 2016:2019) {
temp <- SPI %>%
  filter(date==reference_year) %>%
  select( SPI.OVRL.SCR, contains('SPI.D1'),contains('SPI.D2'),contains('SPI.D3'),contains('SPI.D4')) %>%
  my_skim() %>%
  yank('numeric') %>%
  rename(Series=skim_variable) %>%
  left_join(metadata) %>% #add labels to variables
  select(Series, 'Indicator Name', mean, sd,p0,p100, N)

assign(paste("spi_unweighted_sumstats",reference_year,sep="_"), temp)

}

#save these to an excel file for easy sharing
write.xlsx(
  spi_unweighted_sumstats_2016,
  file=paste(out_dir, 'summary_statistics_SPI.xlsx', sep="/"),
  sheetName='spi_unweighted_sumstats_2016',
  append=T
) 

write.xlsx(
  spi_unweighted_sumstats_2017,
  file=paste(out_dir, 'summary_statistics_SPI.xlsx', sep="/"),
  sheetName='spi_unweighted_sumstats_2017',
  append=T
)      

write.xlsx(
  spi_unweighted_sumstats_2018,
  file=paste(out_dir, 'summary_statistics_SPI.xlsx', sep="/"),
  sheetName='spi_unweighted_sumstats_2018',
  append=T
) 
write.xlsx(
  spi_unweighted_sumstats_2019,
  file=paste(out_dir, 'summary_statistics_SPI.xlsx', sep="/"),
  sheetName='spi_unweighted_sumstats_2019',
  append=T
) 

#create grouping variable for table
group <- spi_unweighted_sumstats_2019 %>%
  mutate(group=case_when(
    grepl('OVRL', Series) ~ "Overall",
    grepl('D1', Series) ~ "Dimension 1",
    grepl('D2', Series) ~ "Dimension 2",
    grepl('D3', Series) ~ "Dimension 3",
    grepl('D4', Series) ~ "Dimension 4"
  ))

kable(spi_unweighted_sumstats_2019,
      digits=2,
      col.names=c('Series', 'Description', 'Mean','SD','Min','Max', 'N'),
      caption="Statistical Performance Indicator Unweighted Summary Statistics for 2019") %>%
  kable_styling(bootstrap_options = c('striped','hover')) %>%
  pack_rows(index=table(fct_inorder(group$group)))
      
      



```

## Regional Means (Unweighted)

```{r regional_unweighted, echo=FALSE}

for (reference_year in 2016:2019) {
df_overall <- SPI %>%
  filter(date==reference_year) 
#produce by region
        sumstats<- df_overall %>%
            group_by(region) %>%
            filter(!is.na(region)) %>%
            select(region, SPI.OVRL.SCR, contains('SPI.D1'),contains('SPI.D2'),contains('SPI.D3'),contains('SPI.D4'), population) %>%
            summarise_all(~round(100*mean(., na.rm=T),2)) %>%
            select(-population) %>%
            mutate(SPI.OVRL.SCR=round(SPI.OVRL.SCR/100,2),
                   SPI.D1.MSC=round(SPI.D1.MSC/100,2),
                   SPI.D2.CS=round(SPI.D2.CS/100,2),
                   SPI.D3.AKI=round(SPI.D3.AKI/100,2),
                   SPI.D4.DPO=round(SPI.D4.DPO/100,2)) 
        
        #produce global number
        sumstats_gl<- df_overall %>%
            mutate(region='Global') %>%
            group_by(region) %>%
            select(region, SPI.OVRL.SCR, contains('SPI.D1'),contains('SPI.D2'),contains('SPI.D3'),contains('SPI.D4'), population) %>%
            summarise_all(~round(100*mean(., na.rm=T),2)) %>%
            select(-population) %>%
            mutate(SPI.OVRL.SCR=round(SPI.OVRL.SCR/100,2),
                   SPI.D1.MSC=round(SPI.D1.MSC/100,2),
                   SPI.D2.CS=round(SPI.D2.CS/100,2),
                   SPI.D3.AKI=round(SPI.D3.AKI/100,2),
                   SPI.D4.DPO=round(SPI.D4.DPO/100,2))
        
        
        #transpose data
        sumstats_df_long <-sumstats_gl %>%
            bind_rows(sumstats) 
        
        sumstats_df <- as.data.frame(t(sumstats_df_long %>% select(-region)))
        colnames(sumstats_df) = sumstats_df_long$region 
        
        
        sumstats_df <- sumstats_df %>%
            rownames_to_column() %>%
            rename(series=rowname)
        
        
        #create labels df
        metadata_tab2_overall <- metadata %>% 
            janitor::clean_names() %>%
            select(series, indicator_name)
        
        
        #add variable label
        sumstats_df <- sumstats_df %>%
            left_join(metadata_tab2_overall) %>%
            rename(Series=series,
                   Label=indicator_name) %>%
            select(Series, Label, everything())

 assign(paste("spi_unweighted_regional",reference_year,sep="_"), sumstats_df)
       
        }
 #save these to an excel file for easy sharing
write.xlsx(
  spi_unweighted_regional_2016,
  file=paste(out_dir, 'summary_statistics_SPI.xlsx', sep="/"),
  sheetName='spi_unweighted_regional_2016',
  append=T
) 

write.xlsx(
  spi_unweighted_regional_2017,
  file=paste(out_dir, 'summary_statistics_SPI.xlsx', sep="/"),
  sheetName='spi_unweighted_regional_2017',
  append=T
)     

write.xlsx(
  spi_unweighted_regional_2018,
  file=paste(out_dir, 'summary_statistics_SPI.xlsx', sep="/"),
  sheetName='spi_unweighted_regional_2018',
  append=T
)     
write.xlsx(
  spi_unweighted_regional_2019,
  file=paste(out_dir, 'summary_statistics_SPI.xlsx', sep="/"),
  sheetName='spi_unweighted_regional_2019',
  append=T
)     
DT::datatable(spi_unweighted_regional_2019, caption=htmltools::tags$caption(
            style = 'caption-side: top; text-align: left;',
            htmltools::em("Unweighted Mean of SPI Indicators by Region in 2019.  Sub-indicators scaled to be 0-100."))    ,
            extensions = 'Buttons', options=list(
                dom = 'Bfrtip',
                buttons = c('copy', 'csv', 'excel', 'pdf', 'print'),
                pageLength = 60)) 

```

## Income Group Means (Unweighted)

```{r income_unweighted, echo=FALSE}

for (reference_year in 2016:2019) {
df_overall <- SPI %>%
  filter(date==reference_year) 
#produce by income
        sumstats<- df_overall %>%
            group_by(income) %>%
            filter(!is.na(income)) %>%
            select(income, SPI.OVRL.SCR, contains('SPI.D1'),contains('SPI.D2'),contains('SPI.D3'),contains('SPI.D4'), population) %>%
            summarise_all(~round(100*mean(., na.rm=T),2)) %>%
            select(-population) %>%
            mutate(SPI.OVRL.SCR=round(SPI.OVRL.SCR/100,2),
                   SPI.D1.MSC=round(SPI.D1.MSC/100,2),
                   SPI.D2.CS=round(SPI.D2.CS/100,2),
                   SPI.D3.AKI=round(SPI.D3.AKI/100,2),
                   SPI.D4.DPO=round(SPI.D4.DPO/100,2)) 
        
        #produce global number
        sumstats_gl<- df_overall %>%
            mutate(income='Global') %>%
            group_by(income) %>%
            select(income, SPI.OVRL.SCR, contains('SPI.D1'),contains('SPI.D2'),contains('SPI.D3'),contains('SPI.D4'), population) %>%
            summarise_all(~round(100*mean(., na.rm=T),2)) %>%
            select(-population) %>%
            mutate(SPI.OVRL.SCR=round(SPI.OVRL.SCR/100,2),
                   SPI.D1.MSC=round(SPI.D1.MSC/100,2),
                   SPI.D2.CS=round(SPI.D2.CS/100,2),
                   SPI.D3.AKI=round(SPI.D3.AKI/100,2),
                   SPI.D4.DPO=round(SPI.D4.DPO/100,2))
        
        
        #transpose data
        sumstats_df_long <-sumstats_gl %>%
            bind_rows(sumstats) 
        
        sumstats_df <- as.data.frame(t(sumstats_df_long %>% select(-income)))
        colnames(sumstats_df) = sumstats_df_long$income 
        
        
        sumstats_df <- sumstats_df %>%
            rownames_to_column() %>%
            rename(series=rowname)
        
        
        #create labels df
        metadata_tab2_overall <- metadata %>% 
            janitor::clean_names() %>%
            select(series, indicator_name)
        
        
        #add variable label
        sumstats_df <- sumstats_df %>%
            left_join(metadata_tab2_overall) %>%
            rename(Series=series,
                   Label=indicator_name) %>%
            select(Series, Label, everything())

 assign(paste("spi_unweighted_income",reference_year,sep="_"), sumstats_df)
       
        }
 #save these to an excel file for easy sharing
write.xlsx(
  spi_unweighted_income_2016,
  file=paste(out_dir, 'summary_statistics_SPI.xlsx', sep="/"),
  sheetName='spi_unweighted_income_2016',
  append=T
) 

write.xlsx(
  spi_unweighted_income_2017,
  file=paste(out_dir, 'summary_statistics_SPI.xlsx', sep="/"),
  sheetName='spi_unweighted_income_2017',
  append=T
)     

write.xlsx(
  spi_unweighted_income_2018,
  file=paste(out_dir, 'summary_statistics_SPI.xlsx', sep="/"),
  sheetName='spi_unweighted_income_2018',
  append=T
)     
write.xlsx(
  spi_unweighted_income_2019,
  file=paste(out_dir, 'summary_statistics_SPI.xlsx', sep="/"),
  sheetName='spi_unweighted_income_2019',
  append=T
)     
DT::datatable(spi_unweighted_income_2019, caption=htmltools::tags$caption(
            style = 'caption-side: top; text-align: left;',
            htmltools::em("Unweighted Mean of SPI Indicators by Income in 2019.  Sub-indicators scaled to be 0-100."))    ,
            extensions = 'Buttons', options=list(
                dom = 'Bfrtip',
                buttons = c('copy', 'csv', 'excel', 'pdf', 'print'),
                pageLength = 60)) 

```


# Appendix

## Summary Statistics 2016

Below we show summary statistics for the year 2016.

```{r sumstats_2016, echo=FALSE}

kable(spi_weighted_sumstats_2016,
      digits=2,
      col.names=c('Series', 'Description', 'Mean','SD','Min','Max', 'N'),
      caption="Statistical Performance Indicator Population Weighted Summary Statistics for 2016") %>%
  kable_styling(bootstrap_options = c('striped','hover')) %>%
  pack_rows(index=table(fct_inorder(group$group)))

DT::datatable(spi_weighted_regional_2016, caption=htmltools::tags$caption(
            style = 'caption-side: top; text-align: left;',
            htmltools::em("Population Weighted Mean of SPI Indicators by Region in 2016.  Sub-indicators scaled to be 0-100."))    ,
            extensions = 'Buttons', options=list(
                dom = 'Bfrtip',
                buttons = c('copy', 'csv', 'excel', 'pdf', 'print'),
                pageLength = 60)) 

DT::datatable(spi_weighted_income_2016, caption=htmltools::tags$caption(
            style = 'caption-side: top; text-align: left;',
            htmltools::em("Population Weighted Mean of SPI Indicators by Income in 2016.  Sub-indicators scaled to be 0-100."))    ,
            extensions = 'Buttons', options=list(
                dom = 'Bfrtip',
                buttons = c('copy', 'csv', 'excel', 'pdf', 'print'),
                pageLength = 60)) 

kable(spi_unweighted_sumstats_2016,
      digits=2,
      col.names=c('Series', 'Description', 'Mean','SD','Min','Max', 'N'),
      caption="Statistical Performance Indicator Unweighted Summary Statistics for 2016") %>%
  kable_styling(bootstrap_options = c('striped','hover')) %>%
  pack_rows(index=table(fct_inorder(group$group)))


        
DT::datatable(spi_unweighted_regional_2016, caption=htmltools::tags$caption(
            style = 'caption-side: top; text-align: left;',
            htmltools::em("Unweighted Mean of SPI Indicators by Region in 2016.  Sub-indicators scaled to be 0-100."))    ,
            extensions = 'Buttons', options=list(
                dom = 'Bfrtip',
                buttons = c('copy', 'csv', 'excel', 'pdf', 'print'),
                pageLength = 60))         

DT::datatable(spi_unweighted_income_2016, caption=htmltools::tags$caption(
            style = 'caption-side: top; text-align: left;',
            htmltools::em("Unweighted Mean of SPI Indicators by Income in 2016.  Sub-indicators scaled to be 0-100."))    ,
            extensions = 'Buttons', options=list(
                dom = 'Bfrtip',
                buttons = c('copy', 'csv', 'excel', 'pdf', 'print'),
                pageLength = 60)) 

```



## Summary Statistics 2017

Below we show summary statistics for the year 2017.

```{r sumstats_2017, echo=FALSE}
kable(spi_weighted_sumstats_2017,
      digits=2,
      col.names=c('Series', 'Description', 'Mean','SD','Min','Max', 'N'),
      caption="Statistical Performance Indicator Population Weighted Summary Statistics for 2017") %>%
  kable_styling(bootstrap_options = c('striped','hover')) %>%
  pack_rows(index=table(fct_inorder(group$group)))

DT::datatable(spi_weighted_regional_2017, caption=htmltools::tags$caption(
            style = 'caption-side: top; text-align: left;',
            htmltools::em("Population Weighted Mean of SPI Indicators by Region in 2017.  Sub-indicators scaled to be 0-100."))    ,
            extensions = 'Buttons', options=list(
                dom = 'Bfrtip',
                buttons = c('copy', 'csv', 'excel', 'pdf', 'print'),
                pageLength = 60)) 

DT::datatable(spi_weighted_income_2017, caption=htmltools::tags$caption(
            style = 'caption-side: top; text-align: left;',
            htmltools::em("Population Weighted Mean of SPI Indicators by Income in 2017.  Sub-indicators scaled to be 0-100."))    ,
            extensions = 'Buttons', options=list(
                dom = 'Bfrtip',
                buttons = c('copy', 'csv', 'excel', 'pdf', 'print'),
                pageLength = 60)) 

kable(spi_unweighted_sumstats_2017,
      digits=2,
      col.names=c('Series', 'Description', 'Mean','SD','Min','Max', 'N'),
      caption="Statistical Performance Indicator Unweighted Summary Statistics for 2017") %>%
  kable_styling(bootstrap_options = c('striped','hover')) %>%
  pack_rows(index=table(fct_inorder(group$group)))

DT::datatable(spi_unweighted_regional_2017, caption=htmltools::tags$caption(
            style = 'caption-side: top; text-align: left;',
            htmltools::em("Unweighted Mean of SPI Indicators by Region in 2017.  Sub-indicators scaled to be 0-100."))    ,
            extensions = 'Buttons', options=list(
                dom = 'Bfrtip',
                buttons = c('copy', 'csv', 'excel', 'pdf', 'print'),
                pageLength = 60))    

DT::datatable(spi_unweighted_income_2017, caption=htmltools::tags$caption(
            style = 'caption-side: top; text-align: left;',
            htmltools::em("Unweighted Mean of SPI Indicators by Income in 2017.  Sub-indicators scaled to be 0-100."))    ,
            extensions = 'Buttons', options=list(
                dom = 'Bfrtip',
                buttons = c('copy', 'csv', 'excel', 'pdf', 'print'),
                pageLength = 60))  
```



## Summary Statistics 2018

Below we show summary statistics for the year 2018.

```{r sumstats_2018, echo=FALSE}
kable(spi_weighted_sumstats_2018,
      digits=2,
      col.names=c('Series', 'Description', 'Mean','SD','Min','Max', 'N'),
      caption="Statistical Performance Indicator Population Weighted Summary Statistics for 2018") %>%
  kable_styling(bootstrap_options = c('striped','hover')) %>%
  pack_rows(index=table(fct_inorder(group$group)))

DT::datatable(spi_weighted_regional_2018, caption=htmltools::tags$caption(
            style = 'caption-side: top; text-align: left;',
            htmltools::em("Population Weighted Mean of SPI Indicators by Region in 2018.  Sub-indicators scaled to be 0-100."))    ,
            extensions = 'Buttons', options=list(
                dom = 'Bfrtip',
                buttons = c('copy', 'csv', 'excel', 'pdf', 'print'),
                pageLength = 60)) 

DT::datatable(spi_weighted_income_2018, caption=htmltools::tags$caption(
            style = 'caption-side: top; text-align: left;',
            htmltools::em("Population Weighted Mean of SPI Indicators by Income in 2018.  Sub-indicators scaled to be 0-100."))    ,
            extensions = 'Buttons', options=list(
                dom = 'Bfrtip',
                buttons = c('copy', 'csv', 'excel', 'pdf', 'print'),
                pageLength = 60)) 

kable(spi_unweighted_sumstats_2018,
      digits=2,
      col.names=c('Series', 'Description', 'Mean','SD','Min','Max', 'N'),
      caption="Statistical Performance Indicator Unweighted Summary Statistics for 2018") %>%
  kable_styling(bootstrap_options = c('striped','hover')) %>%
  pack_rows(index=table(fct_inorder(group$group)))

DT::datatable(spi_unweighted_regional_2018, caption=htmltools::tags$caption(
            style = 'caption-side: top; text-align: left;',
            htmltools::em("Unweighted Mean of SPI Indicators by Region in 2018.  Sub-indicators scaled to be 0-100."))    ,
            extensions = 'Buttons', options=list(
                dom = 'Bfrtip',
                buttons = c('copy', 'csv', 'excel', 'pdf', 'print'),
                pageLength = 60))            

DT::datatable(spi_unweighted_income_2018, caption=htmltools::tags$caption(
            style = 'caption-side: top; text-align: left;',
            htmltools::em("Unweighted Mean of SPI Indicators by Income in 2018.  Sub-indicators scaled to be 0-100."))    ,
            extensions = 'Buttons', options=list(
                dom = 'Bfrtip',
                buttons = c('copy', 'csv', 'excel', 'pdf', 'print'),
                pageLength = 60))   
```
